% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{abBoxPlot}
\alias{abBoxPlot}
\title{abundance boxplot plot}
\usage{
abBoxPlot(
  object,
  cluster_idents,
  sample,
  cluster_order,
  group_by,
  group_order,
  color,
  width = 9,
  height = ceiling(length(unique(object@meta.data[[cluster_idents]]))/4) * 3,
  paired = FALSE,
  number_of_tests,
  dir_output = "."
)
}
\arguments{
\item{object}{Seurat object}

\item{cluster_idents}{variable in meta data with cluster names}

\item{sample}{variable in meta data for each sample}

\item{cluster_order}{vector determining the order of the clusters}

\item{group_by}{variable in meta data that categorize samples in groups}

\item{group_order}{vector determining the order of the samples}

\item{color}{color palette}

\item{width}{width of output plot (default: 9)}

\item{height}{height of output plot (default: length of cluster_idents divided by four, ceiling, times three)}

\item{paired}{logical indicating whether you want a paired test (default FALSE)}

\item{number_of_tests}{number of tests to be performed}

\item{dir_output}{directory to save the output plot (default: ".")}
}
\value{
save abundance box plot
}
\description{
create and save an abundance boxplot in the folder \code{abundance}
}
\examples{
library(Seurat)
set.seed(123)
pbmc_small$cluster <- sample(c("Cluster1", "Cluster2"), ncol(pbmc_small), replace = TRUE)
pbmc_small$sample <- sample(c("CSF_P01", "CSF_P02", "CSF_P03", "CSF_P04"), ncol(pbmc_small), replace = TRUE)
lookup <-
  data.frame(
    sample = c("CSF_P01", "CSF_P02", "CSF_P03", "CSF_P04"),
    AIE_type = c(rep("control", 2), rep("CASPR2", 2))
  )
pbmc_small@meta.data <-
  pbmc_small@meta.data |>
  tibble::rownames_to_column("barcode") |>
  dplyr::left_join(lookup, by = "sample") |>
  tibble::column_to_rownames("barcode")
abBoxPlot(
  object = pbmc_small,
  cluster_idents = "cluster",
  sample = "sample",
  cluster_order = c("Cluster1", "Cluster2"),
  group_by = "AIE_type",
  group_order = c("control", "CASPR2"),
  color = c("control" = "blue", "CASPR2" = "red"),
  width = 9,
  height = 6,
  paired = FALSE,
  number_of_tests = 3,
  dir_output = "."
)
unlink("boxplot_cluster_pbmc_small_AIE_type.pdf")
}
