% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{abVolPlot}
\alias{abVolPlot}
\title{abundance volcano plot}
\usage{
abVolPlot(
  object,
  cluster_idents,
  sample,
  cluster_order,
  group_by,
  group1,
  group2,
  color,
  width = 5,
  height = 5,
  min_cells = 10,
  paired = FALSE
)
}
\arguments{
\item{object}{Seurat object}

\item{cluster_idents}{variable in meta data with cluster names}

\item{sample}{variable in meta data for each sample}

\item{cluster_order}{vector determining the order of the clusters}

\item{group_by}{variable in meta data that categorize samples in groups}

\item{group1}{first group (nominator)}

\item{group2}{second group (denominator)}

\item{color}{color palette}

\item{width}{width of output plot (default: 5)}

\item{height}{height of output plot (default: 5)}

\item{min_cells}{remove all clusters that have less than minimal amount of cells (default = 10)}

\item{paired}{logical indicating whether you want a paired test (default FALSE)}
}
\value{
save volcano abundance plot to folder \verb{/results/abundance}
}
\description{
create and save an abundance volcano bar plot in the folder \code{abundance}
}
\examples{
\dontrun{
abVolPlot(object = aie_sct,
}
}
