% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{as.CellDataSet}
\alias{as.CellDataSet}
\alias{as.CellDataSet.SingleCellExperiment}
\title{as.CellDataSet}
\usage{
as.CellDataSet(x, expressionFamily = NULL, gene_short_name = "symbol")

\method{as.CellDataSet}{SingleCellExperiment}(x, expressionFamily = NULL, gene_short_name = "symbol")
}
\arguments{
\item{x}{object to convert.}

\item{expressionFamily}{expression family function to use (default: negbinomial.size)}

\item{gene_short_name}{column to use to set gene_short_name}
}
\value{
a CellDataSet object.
}
\description{
Convert several data types into monocle::CellDataSet object.
}
