% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDiffusionMap.R
\name{calculateDiffusionMap}
\alias{calculateDiffusionMap}
\alias{calculateDiffusionMap,ANY-method}
\alias{calculateDiffusionMap,SummarizedExperiment-method}
\alias{calculateDiffusionMap,SingleCellExperiment-method}
\alias{runDiffusionMap}
\title{Create a diffusion map of cells}
\usage{
calculateDiffusionMap(x, ...)

\S4method{calculateDiffusionMap}{ANY}(x, ncomponents = 20L, ..., tips = NULL, w_width = 0.1)

\S4method{calculateDiffusionMap}{SummarizedExperiment}(x, ..., assay.type = "logcounts")

\S4method{calculateDiffusionMap}{SingleCellExperiment}(
  x,
  ...,
  assay.type = "logcounts",
  dimred = "PCA",
  n_dimred = NULL
)

runDiffusionMap(x, ..., altexp = NULL, name = "DiffusionMap")
}
\arguments{
\item{x}{For \code{calculateDiffusionMap}, a numeric matrix of log-expression
values where rows are features and columns are cells. Alternatively, a
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment} or
\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment} object
containing such a matrix.

For \code{runDiffusionMap}, a
\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment} object.}

\item{...}{For the \code{calculateDiffusionMap} generic, additional arguments to
pass to specific methods.

For the \link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment} or
\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment} methods,
additional arguments to pass to the ANY method.

For the ANY methods, additional arguments to pass to \link[destiny:DiffusionMap-class]{destiny::DiffusionMap}
\itemize{
\item \code{sigma}: Diffusion scale parameter of the Gaussian kernel. One of
\code{'local'}, \code{'global'}, a (\link[base]{numeric}) global
sigma or a \link[destiny:Sigmas-class]{Sigmas} object.  When choosing
\code{'global'}, a global sigma will be calculated using
\link[destiny:find_sigmas]{find_sigmas}.  (Optional. default:
\code{'local'}) A larger sigma might be necessary if the
eigenvalues can not be found because of a singularity in the
matrix.
\item \code{k}: Number of nearest neighbors to consider (default: a guess betweeen
100 and \eqn{n - 1}. See \link[destiny:find_dm_k]{find_dm_k}).
\item \code{density_norm}: logical. If TRUE, use density normalisation.
\item \code{n_local}: If \code{sigma == 'local'}, the \code{n_local}th nearest
neighbor(s) determine(s) the local sigma.
\item \code{censor_val}: Value regarded as uncertain. Either a single value or one
for every dimension (Optional, default: censor_val).
\item \code{censor_range}: Uncertainity range for censoring (Optional, default:
none).  A length-2-vector of certainty range start and end.
TODO: also allow \eqn{ 2 \times G } matrix.
\item \code{missing_range}: Whole data range for missing value model. Has to be
specified if NAs are in the data.
\item \code{knn_params}: Parameters passed to \link[destiny:knn]{find_knn}.
\item \code{verbose}: Show a progressbar and other progress information (default: do
it if censoring is enabled).
}

For \code{runDiffusionMap}, additional arguments to pass to
\code{calculateDiffusionMap}.}

\item{ncomponents}{Numeric scalar indicating the number of principal components to obtain.}

\item{tips}{The cell index/indices from which to calculate the DPT(s) (integer of length 1-3)}

\item{w_width}{Window width to use for deciding the branch cutoff}

\item{assay.type}{Integer scalar or string indicating which assay of \code{x} contains the expression values.}

\item{dimred}{String or integer scalar specifying the existing dimensionality reduction results to use.}

\item{n_dimred}{Integer scalar or vector specifying the dimensions to use if \code{dimred} is specified.}

\item{altexp}{String or integer scalar specifying an alternative experiment containing the input data.}

\item{name}{String specifying the name to be used to store the result in the \code{\link{reducedDims}} of the output.}
}
\value{
For \code{calculateDiffusionMap}, a numeric matrix of coordinates for each cell
(row) in each of \code{ncomponents} DiffusionMap (column).

For \code{runDiffusionMap}, a
\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment} object is
returned containing this matrix in \code{\link{reducedDims}(..., name)}.

In both cases, the attributes of the DiffusionMap coordinate matrix contain
the \code{DPT}: the raw output of \link[destiny:DPT]{DPT}.
}
\description{
Create a diffusion map of cells
}
