% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_helper.R
\name{calculatePropellerPermFDP}
\alias{calculatePropellerPermFDP}
\title{Calculate permFDP-adjusted threshold for propeller results}
\usage{
calculatePropellerPermFDP(
  prop_data,
  seu_obj,
  sample_col,
  cluster_col,
  meta_col,
  cond1,
  cond2,
  fdr_threshold,
  n_perms = 1000
)
}
\arguments{
\item{prop_data}{A dataframe containing propeller results with P.Value column}

\item{seu_obj}{A Seurat object subset to the two conditions being compared}

\item{sample_col}{A character representing the name of the sample column}

\item{cluster_col}{A character representing the name of the cluster column}

\item{meta_col}{A character representing the name of the condition column}

\item{cond1}{A character representing the first condition}

\item{cond2}{A character representing the second condition}

\item{fdr_threshold}{The FDR threshold to use for permFDP adjustment}

\item{n_perms}{Number of permutations for permFDP adjustment (default: 1000)}
}
\value{
A dataframe with added permFDP_sig and permFDP_threshold columns
}
\description{
Internal helper function to calculate permFDP-adjusted significance threshold for propeller differential abundance testing
}
\keyword{internal}
