% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chan_mnn.R
\name{chan_mnn}
\alias{chan_mnn}
\title{Fast mutual nearest neighbors correction with \code{scran.chan}}
\usage{
chan_mnn(
  x,
  batch = NULL,
  assay.type = "counts",
  subset.row = NULL,
  d = 50L,
  k = 20L,
  scale = FALSE,
  batch.mode = NULL,
  batch.method = NULL,
  ...,
  approximate = TRUE,
  threads = 1L
)
}
\arguments{
\item{x}{A single
\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment} object.}

\item{batch}{Vector or factor of length equal to the number of cells, specifying the batch of origin for each cell.
Alternatively \code{NULL} if all cells belong to the same batch.}

\item{assay.type}{A string specifying which assay values contains the counts.}

\item{subset.row}{Integer, logical or character vector specifying which
features to use in the PCA (e.g., highly variable genes). If NULL, all
features in x are used.}

\item{d}{Integer scalar specifying the number of top PCs to obtain.}

\item{k}{Integer scalar specifying the number of neighbors to use when
identifying MNN pairs.}

\item{scale}{Logical scalar indicating whether to scale rows to unit variance.}

\item{batch.mode}{String indicating how \code{batch} should be handled when centering the size factors.
If \code{"lowest"}, we downscale all batches to the coverage of the lowest batch.
If \code{"perblock"}, we scale each batch to a mean of 1.}

\item{batch.method}{String indicating how \code{batch} should be handled (if it is supplied).
\code{"block"} is equivalent to linear regression on \code{x} prior to PCA,
while \code{"weight"} will only weight each batch so that they contribute equally to the PCA.}

\item{...}{
  Arguments passed on to \code{\link[scran.chan:mnnCorrect.chan]{scran.chan::mnnCorrect.chan}}
  \describe{
    \item{\code{nmads}}{Numeric scalar specifying the number of MADs to use when computing correction vectors.}
    \item{\code{mass.cap}}{Integer scalar specifying the cap on the number of observations to use for center of mass calculations.
A value of 100,000 may be appropriate for speeding up correction of very large datasets.
If this is set to NULL, no cap is used.}
    \item{\code{order}}{Vector containing levels of \code{batch} in the desired merge order.
If \code{NULL}, a suitable merge order is automatically determined.}
    \item{\code{reference.policy}}{String specifying the policy to use to choose the first reference batch.
This can be based on the largest batch (\code{"max-size"}), the most variable batch (\code{"max-variance"}), 
some combination of those two (\code{"max-rss"}) or the first specified input (\code{"input"}).
Only used for automatic merges, i.e., when \code{order=NULL}.}
  }}

\item{approximate}{Logical scalar specifying whether to perform an
approximate neighbor search.}

\item{threads}{Integer scalar specifying the number of threads to use.}
}
\description{
Fast mutual nearest neighbors correction with \code{scran.chan}
}
\seealso{
\itemize{
\item \link[scran.chan:logNormCounts.chan]{logNormCounts.chan}
\item \link[scran.chan:runPCA.chan]{runPCA.chan}
\item \link[scran.chan:mnnCorrect.chan]{mnnCorrect.chan}
}
}
