% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster_cells}
\alias{cluster_cells}
\alias{cluster_cells.SingleCellExperiment}
\title{cluster_cells}
\usage{
cluster_cells(x, method = "kmeans", ...)

\method{cluster_cells}{SingleCellExperiment}(x, method = "kmeans",
  ncluster = NULL, assay.name = "logcounts", coord.name = "PCA",
  column.name = "cluster", hclust.method = "ward.D",
  dist.method = "euclidean", resolution = 1, algorithm = 3, ...)
}
\arguments{
\item{x}{object with data matrix.}

\item{method}{type of dimensionality reduction.}

\item{...}{arguments passed down to methods.}

\item{ncluster}{number of clusters.}

\item{assay.name}{name of assay slot.}

\item{coord.name}{name of reducedDim slot.}

\item{column.name}{name of column to store cluster class.}

\item{hclust.method}{method for hierarchical clustering (default: ward.D).}

\item{dist.method}{method for distance calculation (default: euclidean).}

\item{resolution}{parameter for seurat method (see ?Seurat::FindClusters).}

\item{algorithm}{parameter for seurat method (see ?Seurat::FindClusters).}
}
\description{
cluster_cells
}
