% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster_genes}
\alias{cluster_genes}
\alias{cluster_genes.Seurat}
\title{cluster_genes}
\usage{
cluster_genes(x, ...)

\method{cluster_genes}{Seurat}(
  x,
  features = NULL,
  reduction = "pca",
  assay = NULL,
  slot = "data",
  dims = 1:10,
  approx = TRUE,
  algorithm = "leiden",
  resolution = 0.6,
  return.seurat = FALSE,
  ...
)
}
\arguments{
\item{x}{object with genes to clusters.}

\item{...}{arguments passed down to methods.}

\item{features}{which features to cluster.}

\item{reduction}{what reduction embeddings to use.}

\item{assay}{assay for Seurat objects.}

\item{slot}{slot for Seurat objects.}

\item{dims}{dimensions to use.}

\item{approx}{whether to use approximate method for PCA.}

\item{algorithm}{algorithm for clustering.}

\item{resolution}{resolution for clustering.}

\item{return.seurat}{whether to return a Seurat object.}
}
\description{
Cluster genes into modules using louvain/leiden algorithm
}
