% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{clustifyFun}
\alias{clustifyFun}
\title{clustifyr UMAP heatmap wrap}
\usage{
clustifyFun(
  my_object,
  my_annotations,
  ref_object,
  ref_annotations,
  format,
  ortho = "none",
  filter = FALSE,
  regex = "",
  umap = FALSE,
  width = 10,
  height = 7,
  cluster_rows = TRUE,
  cluster_cols = TRUE,
  cellwidth = 10,
  cellheight = 10,
  fontsize = 10,
  color,
  name
)
}
\arguments{
\item{my_object}{query input object}

\item{my_annotations}{annotations of query input object}

\item{ref_object}{reference object}

\item{ref_annotations}{reference annotations}

\item{format}{format of reference object: possible values are seurat, sc or matrix}

\item{ortho}{convert to orthologues? Allowed values: \code{none}, \code{mouse2human} or \code{human2mouse} (default: \code{none})}

\item{filter}{should the reference be filtered? Allowed values: TRUE, FALSE (default: FALSE)}

\item{regex}{if filter: specify filter term}

\item{umap}{also create UMAP? Allowed values: TRUE, FALSE (default FALSE)}

\item{width}{width of heatmap plot (default: 10)}

\item{height}{height of heatmap plot (default: 7)}

\item{cluster_rows}{cluster rows? (default: true)}

\item{cluster_cols}{cluster columns? (default: true)}

\item{cellwidth}{individual cell width (default: 10)}

\item{cellheight}{individual cell height (default: 10)}

\item{fontsize}{fontsize (default: 10)}

\item{color}{color palette}

\item{name}{fontsize (default: 10)}
}
\value{
save umap/heatmap plot to folder \verb{/results/clustifyr/}
}
\description{
create and save clustifyr UMAP and heatmap
}
\examples{
\dontrun{
clustifyFun(
    my_object = tcells,
    my_annotations = "cluster",
    ref_object = projectil_til,
    ref_annotations = "functional.cluster",
    format = "seurat",
    ortho = "none",
    height = 25,
    width = 10,
    umap = TRUE,
    color = col_vector,
    name = "projectil_til"
)
}
}
