% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{compStat}
\alias{compStat}
\title{significance for boxplot}
\usage{
compStat(x_var, group, data, paired)
}
\arguments{
\item{x_var}{numeric variable (x in the formula \code{x ~ group})}

\item{group}{grouping variable (group in the formula \code{x ~ group})}

\item{data}{data frame containing the variables}

\item{paired}{logical value. Do you want to do perform pair testing?}
}
\value{
data frame with adjusted signficance values
}
\description{
calculcate significance for ggsignif
}
\examples{
set.seed(123)
data <- data.frame(
  sample = c(paste0("CSF_P0", 1:9)),
  type = c(rep("control", 4), rep("AIE", 5)),
  cluster1 = c(runif(4, 0, 1), runif(5, 99, 100)),
  cluster2 = c(runif(4, 0, 70), runif(5, 20, 100))
)
compStat(x_var = c("cluster1", "cluster2"), group = "type", data = data, paired = FALSE)
}
