% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{compute_density}
\alias{compute_density}
\alias{compute_density.default}
\alias{compute_density.SingleCellExperiment}
\alias{compute_density.CellDataSet}
\alias{cell_density}
\title{compute_density}
\usage{
compute_density(x, ...)

\method{compute_density}{default}(x, y, n = 100, ...)

\method{compute_density}{SingleCellExperiment}(x, name = "TSNE", ...)

\method{compute_density}{CellDataSet}(x, name = "A", ...)

cell_density(x, ...)
}
\arguments{
\item{x}{x coordinate of data}

\item{...}{additional arguments passed down to methods.}

\item{y}{y coordinate of data}

\item{n}{Number of grid points in each direction. Can be scalar or a length-2 integer vector.}

\item{name}{name of coordinates for some methods.}
}
\description{
compute_density
}
\note{
I took this implementation from http://slowkow.com/notes/ggplot2-color-by-density/
}
