% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard.R
\name{compute_jaccard}
\alias{compute_jaccard}
\alias{compute_jaccard.SingleCellExperiment}
\alias{compute_jaccard.DataFrame}
\alias{compute_jaccard.data.frame}
\title{compute_jaccard}
\usage{
compute_jaccard(x, ...)

\method{compute_jaccard}{SingleCellExperiment}(x, ...)

\method{compute_jaccard}{DataFrame}(x, ...)

\method{compute_jaccard}{data.frame}(x, col.x, col.y, ...)
}
\arguments{
\item{x}{object to compute jaccard index on.}

\item{...}{}

\item{col.x}{name of column.}

\item{col.y}{name of column.}
}
\description{
compute_jaccard
}
