% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{compute_percentage_variance}
\alias{compute_percentage_variance}
\alias{compute_percentage_variance.Seurat}
\title{compute_percentage_variance}
\usage{
compute_percentage_variance(x, reduction = "pca")

\method{compute_percentage_variance}{Seurat}(x, reduction = "pca")
}
\arguments{
\item{x}{and object with PCA coordintates}

\item{reduction}{name of the reduction to use with Seurat objects.}
}
\description{
compute_percentage_variance
}
