% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purity.R
\name{compute_purity}
\alias{compute_purity}
\alias{compute_purity.SingleCellExperiment}
\alias{compute_purity.DataFrame}
\alias{compute_purity.data.frame}
\title{compute_purity}
\usage{
compute_purity(x, ...)

\method{compute_purity}{SingleCellExperiment}(x, ...)

\method{compute_purity}{DataFrame}(x, ...)

\method{compute_purity}{data.frame}(x, col.x, col.y, ...)
}
\arguments{
\item{x}{object to compute purity on.}

\item{...}{}

\item{col.x}{name of column.}

\item{col.y}{name of column.}
}
\description{
compute_purity
}
