% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{create_training_sets}
\alias{create_training_sets}
\alias{create_training_sets.SingleCellExperiment}
\alias{create_training_sets.Seurat}
\title{create_training_sets}
\usage{
create_training_sets(x, ...)

\method{create_training_sets}{SingleCellExperiment}(x, frac = 0.1, ...)

\method{create_training_sets}{Seurat}(x, frac = 0.1, ...)
}
\arguments{
\item{x}{SingleCellExperiment object.}

\item{...}{arguments passed down to methods (currently unused).}

\item{frac}{fraction of dataset for testing.}
}
\value{
list with two SingleCellExperiment, one with train and other with test data.
}
\description{
Divides a SingleCellExperiment object into train/test subsets.
}
