% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demux.R
\name{demux}
\alias{demux}
\alias{demux.Seurat}
\title{demux}
\usage{
demux(x, n, ...)

\method{demux}{Seurat}(x, n, assay = NULL)
}
\arguments{
\item{x}{object to demux.}

\item{n}{number of hashtags.}

\item{...}{arguments passed down to methods.}
}
\description{
demux
}
