% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demux.R
\name{demux_cutoff}
\alias{demux_cutoff}
\title{demux_cutoff}
\usage{
demux_cutoff(x, cutoff, assay = "HTO", slot = "data")
}
\arguments{
\item{x}{Seurat object.}

\item{cutoff}{vector of cutoffs to apply to each hashtag.}

\item{assay}{assay from Seurat object.}

\item{slot}{slot from Seurat object.}
}
\description{
Hashtag demultiplexing using cutoffs, decided with plot_pairs().
}
