% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimplot.R
\name{dimPlotMisc}
\alias{dimPlotMisc}
\title{Beautify Seurat's DimPlot Function}
\usage{
dimPlotMisc(
  object,
  groupBy,
  style = 1,
  plot.title = NA,
  legend.point.size = 4,
  reduction = "umap",
  legend.position = "bottom",
  label = FALSE,
  label.size = 4,
  point.size = 0.5
)
}
\arguments{
\item{object}{A Seurat object.}

\item{groupBy}{Grouping variable.}

\item{style}{Style of the plot. Defaults to 1.}

\item{plot.title}{Title of the plot. Defaults to NA.}

\item{legend.point.size}{Size of the legend points. Defaults to 4.}

\item{reduction}{Reduction method to use. Defaults to "umap".}

\item{legend.position}{Position of the legend. Defaults to "bottom".}

\item{label}{Logical; whether to label the plot. Defaults to FALSE.}

\item{label.size}{Size of the labels. Defaults to 4.}

\item{point.size}{Size of the points. Defaults to 0.5.}
}
\value{
A ggplot object.
}
\description{
This function enhances the DimPlot function from the Seurat package by adding
additional customization options for plot aesthetics and layout.
}
\examples{
\dontrun{
dimPlotMisc(seurat_object, groupBy = "celltype", style = 1, label = TRUE)
}

}
