% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{dotPlot}
\alias{dotPlot}
\title{nice Seurat dot plot}
\usage{
dotPlot(
  path,
  object,
  par,
  dot_min,
  scale = TRUE,
  ortho = "none",
  width = 10,
  height = 10,
  dir_output = "."
)
}
\arguments{
\item{path}{path to markers.csv}

\item{object}{Seurat object}

\item{par}{column name in markers.csv}

\item{dot_min}{minimal dot size}

\item{scale}{should the values be scaled? (default: TRUE)}

\item{ortho}{convert to orthologues? Allowed values: \code{none}, \code{mouse2human} or \code{human2mouse}}

\item{width}{width of output plot (default: 10)}

\item{height}{height of output plot (default: 10)}

\item{dir_output}{directory to save the output plot (default: ".")}
}
\value{
save dot plot
}
\description{
create and save a nice Seurat dot plot
}
\examples{
library(Seurat)
markers <- data.frame(B = c("MS4A1", "CD79A"))
write.csv(markers, "markers.csv")
dotPlot(
  path = "markers.csv",
  object = pbmc_small,
  par = "B",
  dot_min = 0.1,
  dir_output = "."
)
unlink("markers.csv")
unlink("dp_pbmc_small_B.pdf")
}
