% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DotsHeatmap.R
\name{DotsHeatmap}
\alias{DotsHeatmap}
\alias{draw,DotsHeatmap-method}
\title{Constructor method for DotsHeatmap class}
\usage{
DotsHeatmap(
  matrix,
  matrix_size = NULL,
  col = NULL,
  dots_size = c(0.1, 1),
  dots_size_legend_param = list(),
  slice_border_gp = NULL,
  ...
)

\S4method{draw}{DotsHeatmap}(object, ...)
}
\arguments{
\item{matrix}{A matrix. Either numeric or character. If it is a simple
vector, it will be converted to a one-column matrix. Only used for dot
color.}

\item{matrix_size}{A numeric matrix or vector. If it is a simple vector, it
will be converted to a one-column matrix. Only used for dot size.}

\item{col}{A vector of colors if the color mapping is discrete or a color mapping  function if the matrix is continuous numbers (should be generated by \code{\link[circlize]{colorRamp2}}). If the matrix is continuous, the value can also be a vector of colors so that colors can be interpolated. Pass to \code{\link[ComplexHeatmap]{ColorMapping}}. For more details and examples, please refer to \url{https://jokergoo.github.io/ComplexHeatmap-reference/book/a-single-heatmap.html#colors} .}

\item{dots_size}{A numeric vector of length one or two specifying the the
minimal and maximal relative dots radius compared with the heatmap
\code{min(grid::unit.c(width, height)) / 2L}. If length one, all dots will have
the same size. If length two with different values, the \code{matrix_size} will
be scaled into \code{dots_size}. Default value: \code{c(0.1, 1)}, means the maximal
value in matrix_size will have a radius equal to the heatmap
\code{min(grid::unit.c(width, height)) / 2L} and the minimal value will equal to
\code{0.1 * min(grid::unit.c(width, height)) / 2L}.}

\item{dots_size_legend_param}{Other arguments passed to
\code{\link[ComplexHeatmap:Legend]{Legend()}}. This will be used to define the dots
size legend.}

\item{slice_border_gp}{Graphic parameters for drawing slice rectangles. The
value should be specified by \link[grid:gpar]{gpar} and fill parameter is
always setted to "transparent". If \code{NULL}, no slice border will be drawn.}

\item{...}{Other arguments passed to
\link[ComplexHeatmap:draw-HeatmapList-method]{draw-HeatmapList-method}.}

\item{object}{A \link{DotsHeatmap} object returned by \link{DotsHeatmap}.}
}
\value{
A \link{DotsHeatmap} object
}
\description{
Constructor method for DotsHeatmap class

Draw a Dots Heatmap
}
