% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DotsHeatmap.R
\name{draw,DotsHeatmap-method}
\alias{draw,DotsHeatmap-method}
\title{Draw a Dots Heatmap}
\usage{
\S4method{draw}{DotsHeatmap}(object, ...)
}
\arguments{
\item{object}{A \link{DotsHeatmap} object.}

\item{...}{Arguments passed to
\link[ComplexHeatmap:draw-HeatmapList-method]{Heatmap}.}
}
\description{
These objects are imported from other packages. Follow the links below to
see their documentation.
\describe{
\item{draw}{\code{\link[ComplexHeatmap]{draw}}}
}
}
