% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_helper.R
\name{enrichrRun}
\alias{enrichrRun}
\title{enrichment analysis}
\usage{
enrichrRun(sheet, filename, dbs, fc_thresh = 1, p_thresh = 0.001, remove_rp_mt)
}
\arguments{
\item{sheet}{sheet name in excel file}

\item{filename}{name of deg file file without extension}

\item{dbs}{name of the enrichr libraries}

\item{fc_thresh}{log fc threshold (default 1)}

\item{p_thresh}{p value threshold (default 0.001)}

\item{remove_rp_mt}{remove ribosomal and mitochondrial genes? (boolean value)}
}
\value{
save enrichrment analysis in excel sheet with multiple sheets in the folder \code{results/enrichr}
}
\description{
wrapper function to perform enrichment analysis with enrichr and save results
}
\examples{
\dontrun{enrichrFun(filename = "de_ALZ_Naive_blood", dbs = dbs, fc_thresh = 1, p_thresh = 0.001, sheet = "pDC", remove_rp_mt = TRUE)}
}
