% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.R
\name{expand_column}
\alias{expand_column}
\alias{expand_column.data.frame}
\title{expand_column}
\usage{
expand_column(x, ...)

\method{expand_column}{data.frame}(
  x,
  col.name = NULL,
  out.name = col.name,
  value.name = "value",
  ...
)
}
\arguments{
\item{x}{a data.frame or similar object.}

\item{...}{arguments passed down to methods.}

\item{col.name}{column from the data.frame to expand.}

\item{out.name}{column in the data.frame where to store expanded values.}

\item{value.name}{column in the data.frame where to store logical vector.}
}
\description{
expand_column
}
