% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_dots.R
\name{facet_dots}
\alias{facet_dots}
\alias{facet_dots,ANY-method}
\alias{facet_dots,SingleCellExperiment-method}
\title{Create a facet dot plot of expression values}
\usage{
facet_dots(x, ...)

\S4method{facet_dots}{ANY}(
  x,
  marker_list,
  clusters = NULL,
  cluster2cell = NULL,
  flip = TRUE,
  facet_args = list(scales = "free", space = "free"),
  ...
)

\S4method{facet_dots}{SingleCellExperiment}(
  x,
  marker_list,
  clusters = NULL,
  cluster2cell = NULL,
  flip = TRUE,
  facet_args = list(scales = "free", space = "free"),
  ...
)
}
\arguments{
\item{x}{Currently, only
\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment} object is
supported.}

\item{...}{Other arguments passed to \link[scater:plotDots]{plotDots}.}

\item{marker_list}{A named list contaning the markers for each cell types.
names indicate the cell type label and values indicate the markers for this
cell type.}

\item{clusters}{A factor (or vector coercible into a factor) specifying the
group to which each cell in \code{x} belongs. Alternatively, String specifying
the field of \code{colData(x)} containing the grouping factor if \code{x} is a
\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment}, e.g.}

\item{cluster2cell}{A named character or factor returned by
\code{\link[=annotate_clusters]{annotate_clusters()}}.}

\item{facet_args}{A named list passed to \link[ggplot2:facet_grid]{facet_grid}.}

\item{filp}{A scalar logical indicates whether flipping the plot.}
}
\value{
A ggplot2 object.
}
\description{
Create a dot plot of expression values for a grouping of cells, where the
size and color of each dot represents the proportion of detected expression
values and the average expression, respectively, for each feature in each
group of cells.
}
