% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_cells}
\alias{filter_cells}
\alias{filter_cells.SingleCellExperiment}
\alias{filter_cells.matrix}
\title{filter_cells}
\usage{
filter_cells(x, ...)

\method{filter_cells}{SingleCellExperiment}(x, min.gene = 10,
  min.count = 0, assay.name = "counts", ...)

\method{filter_cells}{matrix}(x, min.gene = 10, min.count = 0, ...)
}
\arguments{
\item{x}{object to filter.}

\item{...}{arguments passed down to methods.}

\item{min.gene}{minimum number of genes with min.count for cell to be preserved.}

\item{min.count}{minimum number of counts to be considered expressed.}

\item{assay.name}{name of assay to obtain matrix of counts.}
}
\description{
filter_cells
}
