% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_helper.R
\name{findMarkersPresto}
\alias{findMarkersPresto}
\title{wrapper function for presto find markers}
\usage{
findMarkersPresto(
  ident1,
  ident2 = NULL,
  object,
  only_pos = FALSE,
  min_pct = 0.1,
  logfc_threshold = 0.25,
  assay = assay
)
}
\arguments{
\item{ident1}{cell population 1}

\item{ident2}{cell population 2, if NULL use all (default: NULL)}

\item{object}{Seurat object}

\item{only_pos}{only return positive markers? (default: FALSE)}

\item{min_pct}{minimum fraction of cells in either two of the populations (default: 0.1)}

\item{logfc_threshold}{minimum x-fold difference (default: 0.25)}

\item{assay}{which assay to use in DE testing (e.g. RNA or SCT)}
}
\value{
data frame with significant DE genes arranged by log2FC
}
\description{
find significant DE genes using presto
}
\examples{
library(Seurat, quietly = TRUE)
findMarkersPresto(ident1 = "0", ident2 = "1", object = pbmc_small, assay = "RNA")
}
