% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify.R
\name{fortify.DataFrame}
\alias{fortify.DataFrame}
\title{fortify.DataFrame}
\usage{
fortify.DataFrame(model, data, ...)
}
\arguments{
\item{model}{model or other R object to convert to data frame}

\item{data}{original dataset, if needed}

\item{...}{other arguments passed to methods}
}
\description{
fortify method for DataFrame objects.
}
