% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{fPlot}
\alias{fPlot}
\title{Seurat feature plot}
\usage{
fPlot(
  path,
  object,
  par,
  reduction,
  width = 16,
  height = ceiling(length(genes_found)/4) * 3,
  order
)
}
\arguments{
\item{path}{path to markers.csv}

\item{object}{Seurat object}

\item{par}{column name in markers.csv}

\item{reduction}{a character string specifying the dimension reduction}

\item{width}{width of output plot (default: 16)}

\item{height}{height of output plot (default: length of genes divided by four, ceiling, times three)}

\item{order}{should the feature plot be ordered in order of expression}
}
\value{
save feature plot to folder \verb{/results/featureplot/}
}
\description{
create and save a nice Seurat feature plot in folder \code{featureplot}
}
\examples{
\dontrun{
fPlot(sc_merge, path = file.path("lookup", "markers.csv"), par = "main", reduction = "umap")
}
}
