% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R
\name{get_assay_summary}
\alias{get_assay_summary}
\alias{get_assay_summary.SingleCellExperiment}
\alias{get_assay_summary.Seurat}
\alias{get_expression}
\title{Compute summary from assay matrix}
\usage{
get_assay_summary(x, ...)

\method{get_assay_summary}{SingleCellExperiment}(x, name,
  assay.name = "logcounts", column = "symbol", sum.fun = mean, ...)

\method{get_assay_summary}{Seurat}(x, name, assay.name = "RNA",
  slot = "counts", column = "symbol", sum.fun = mean, ...)

get_expression(...)
}
\arguments{
\item{x}{object to compute summary.}

\item{...}{arguments passed down to specific methods.}

\item{name}{name of the feature value for which to compute the summary.}

\item{assay.name}{name of the assay slot (defualt: 'logcounts').}

\item{column}{name of the column where to find 'name' (default: 'symbol').}

\item{sum.fun}{summary function (by default mean).}
}
\description{
Compute summary from assay matrix
}
