% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.R
\name{get_coord}
\alias{get_coord}
\alias{get_coord.SingleCellExperiment}
\alias{get_coord.CellDataSet}
\title{get_coord}
\usage{
get_coord(x, name = "tSNE", annotate = TRUE, cols = NULL)

\method{get_coord}{SingleCellExperiment}(x, name = "tSNE",
  annotate = TRUE, cols = NULL)

\method{get_coord}{CellDataSet}(x, name = "A", annotate = TRUE,
  cols = NULL)
}
\arguments{
\item{x}{object to obtain coordinates.}

\item{name}{name of the holder for coordinates.}

\item{annotate}{logical; whether to annotate each coordinate.}

\item{cols}{if not NULL the name of columns to include in annotation.}
}
\value{
A data.frame object.
}
\description{
Extract coordinates (e.g. from dimensionality reduction) from bio datasets,
optionally with annotations to help plotting them with ggplot2 package.
}
