% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.R
\name{get_coord_names}
\alias{get_coord_names}
\alias{get_coord_names.SingleCellExperiment}
\alias{get_coord_names.seurat}
\title{get_coord_names}
\usage{
get_coord_names(x)

\method{get_coord_names}{SingleCellExperiment}(x)

\method{get_coord_names}{seurat}(x)
}
\arguments{
\item{x}{object with coordinates}
}
\description{
get_coord_names
}
