% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_data}
\alias{get_data}
\alias{get_data.Seurat}
\title{get_data}
\usage{
get_data(x, coord.name = NULL, feature.name = NULL, meta.data = TRUE, ...)

\method{get_data}{Seurat}(
  x,
  coord.name = NULL,
  feature.name = NULL,
  meta.data = TRUE,
  assay = NULL,
  slot = "data",
  ...
)
}
\arguments{
\item{x}{object to extract data from.}

\item{coord.name}{name of coordinates.}

\item{feature.name}{name of features.}

\item{meta.data}{whether to include meta.data.}

\item{...}{argument passed down to methods.}
}
\description{
Get miscellaneous data from objects.
}
