% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{get_density}
\alias{get_density}
\alias{get_density.default}
\alias{get_density.SingleCellExperiment}
\alias{get_density.CellDataSet}
\title{get_density}
\usage{
get_density(x, ...)

\method{get_density}{default}(x, y, n = 100)

\method{get_density}{SingleCellExperiment}(x, name = "tSNE")

\method{get_density}{CellDataSet}(x, name = "A")
}
\arguments{
\item{x}{x coordinate of data}

\item{y}{y coordinate of data}

\item{n}{Number of grid points in each direction. Can be scalar or a length-2 integer vector.}
}
\description{
get_density
}
\note{
I took this implementation from http://slowkow.com/notes/ggplot2-color-by-density/
}
