% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_rowdata}
\alias{get_rowdata}
\alias{get_rowdata.SingleCellExperiment}
\alias{get_rowdata.seurat}
\alias{get_rowdata.Seurat}
\title{get_rowdata}
\usage{
get_rowdata(x, ...)

\method{get_rowdata}{SingleCellExperiment}(x, ...)

\method{get_rowdata}{seurat}(x, ...)

\method{get_rowdata}{Seurat}(x, assay = NULL, ...)
}
\arguments{
\item{x}{object.}

\item{...}{arguments passed down to methods.}

\item{assay}{name of assay.}
}
\description{
get_rowdata
}
