% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helper.R
\name{lss}
\alias{lss}
\title{list objects from the environment in a pretty way}
\usage{
lss()
}
\value{
tibble with environment objects arranged by size
}
\description{
list objects from the environment, changes to human-readable units and arranges them by size.
Taken from https://stackoverflow.com/a/65949488.
}
\examples{
set.seed(123)
x <- matrix(rnorm(100), 10, 10)
y <- matrix(rnorm(100), 10, 10)
lss()
}
