% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfeatures.R
\name{make_gg}
\alias{make_gg}
\alias{ggfeatures}
\alias{ggcells}
\title{Create a ggplot from a SingleCellExperiment}
\usage{
ggfeatures(
  x,
  mapping = aes(),
  cells = NULL,
  ...,
  assay.type = "logcounts",
  extract_mapping = TRUE
)

ggcells(
  x,
  mapping = aes(),
  features = NULL,
  ...,
  assay.type = "logcounts",
  extract_mapping = TRUE
)
}
\arguments{
\item{x}{A \linkS4class{SingleCellExperiment} object.
This is expected to have non-\code{NULL} row names.}

\item{mapping}{Default list of aesthetic mappings to use for plot.
If not specified, must be supplied in each layer added to the plot.}

\item{cells}{Character vector specifying the features for which to extract
expression profiles across cells. Alternative, A named list contaning
the cell groups, where names indicate the cell group and values indicate
the cells for this group. In this way, data will be transformed into
long-format and a column \code{.features_group} will be added into final data.}

\item{...}{
  Arguments passed on to \code{\link[scuttle:makePerCellDF]{scuttle::makePerCellDF}}
  \describe{
    \item{\code{use.coldata}}{Logical scalar indicating whether column metadata of \code{x} should be included.
Alternatively, a character or integer vector specifying the column metadata fields to use.}
    \item{\code{use.dimred}}{Logical scalar indicating whether data should be extracted for dimensionality reduction results in \code{x}.
Alternatively, a character or integer vector specifying the dimensionality reduction results to use.}
    \item{\code{use.altexps}}{Logical scalar indicating whether (meta)data should be extracted for alternative experiments in \code{x}.
Alternatively, a character or integer vector specifying the alternative experiments to use.}
    \item{\code{prefix.altexps}}{Logical scalar indicating whether \code{\link{altExp}}-derived fields should be prefixed with the name of the alternative Experiment.}
    \item{\code{check.names}}{Logical scalar indicating whether column names of the output should be made syntactically valid and unique.}
    \item{\code{swap.rownames}}{String specifying the \code{\link{rowData}} column containing the \code{features}.
If \code{NULL}, \code{rownames(x)} is used.}
  }}

\item{assay.type}{String or integer scalar indicating the assay to use to obtain expression values.
Must refer to a matrix-like object with integer or numeric values.}

\item{extract_mapping}{Logical scalar indicating whether \code{features} or \code{cells} should be automatically expanded to include variables referenced in \code{mapping}.}

\item{features}{Character vector specifying the features for which to extract
expression profiles across cells. May also include features in alternative
Experiments if permitted by use.altexps. Alternative, A named list contaning
the marker groups, where names indicate the marker group and values indicate
the markers for this group. In this way, data will be transformed into
long-format and a column \code{.cells_group} will be added into final data.}
}
\description{
\code{ggcells} treats cells as the data values so users can reference row names of
\code{x} (if provided in \code{features}), column metadata variables and dimensionality
reduction results. They can also reference row names and metadata variables
for alternative Experiments.
}
\details{
\code{ggfeatures} treats features as the data values so users can reference column
names of \code{x} (if provided in \code{cells}) and row metadata variables.
}
