% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-manual_marker_sets.R
\name{manual_cms}
\alias{manual_cms}
\alias{get_markers}
\alias{print.manual_cms}
\alias{print.marker_set}
\title{Cell markers sets collected from research articles}
\usage{
get_markers(name)

manual_cms()

\method{print}{manual_cms}(x, ...)

\method{print}{marker_set}(x, ...)
}
\arguments{
\item{name}{The name of the marker set to extract. \code{name} usually is the
surname of the first author followed by the years of the published article
connected with "_".}

\item{x}{An \code{manual_cms} or \code{marker_set} object.}

\item{...}{Not used currently}
}
\description{
We manullay collected some cell marker sets from research articles, these
funciton juse provide some methods to operate with the manual cell marker
sets (manual_cms).

\code{manual_cms} function just returns the database with a class of
\code{manual_cell_marker_datasets} which has a specific \code{print} method.

\code{get_markers} function extract the specific marker set from the datasets,
which return a \code{marker_set} class object.
}
