% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{ModulePlot}
\alias{ModulePlot}
\title{module plot}
\usage{
ModulePlot(x_var, module, object, color)
}
\arguments{
\item{x_var}{variable in meta data which represents the x-axis}

\item{module}{variable in meta data of the module score}

\item{object}{Seurat object}

\item{color}{color palette}
}
\value{
plot module plot
}
\description{
create combined violin and boxplot of the module score
}
\examples{
\dontrun{ModulePlot(object = aie_csf, x_var = "AIE_type", module = "TCRVG1", color = my_cols)}
}
