% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{pHeatmap}
\alias{pHeatmap}
\title{nice pheatmap}
\usage{
pHeatmap(
  matrix,
  scale = "none",
  height = ceiling(nrow(matrix)/3),
  width = ceiling(ncol(matrix)/2),
  cellwidth = 10,
  cellheight = 10,
  treeheight_row = 10,
  treeheight_col = 10,
  fontsize = 10,
  cluster_rows = TRUE,
  cluster_cols = TRUE,
  annotation_row = NA,
  dir_output = "."
)
}
\arguments{
\item{matrix}{numeric matrix of the values to be plotted}

\item{scale}{should the values be centered and scaled in row, column direction? Allowed: \code{row}, \code{column}, \code{none} (default: \code{none})}

\item{height}{height of output plot (default: number of rows divided by three)}

\item{width}{width of output plot (default: number of columns divided by two)}

\item{cellwidth}{individual cell width (default: 10)}

\item{cellheight}{individual cell height (default: 10)}

\item{treeheight_row}{height of a tree for rows (default: 10)}

\item{treeheight_col}{height of a tree for columns (default: 10)}

\item{fontsize}{fontsize (default: 10)}

\item{cluster_rows}{cluster rows? (default: true)}

\item{cluster_cols}{cluster columns? (default: true)}

\item{annotation_row}{data frame that contains the annotations. Rows in the data and in the annotation are matched using row names. (default: NA)}

\item{dir_output}{directory to save the output plot (default: ".")}
}
\value{
save heatmap to folder
}
\description{
create and save a nice pheatmap in the folder \code{heatmap} using color breaks
}
\examples{
matrix <- matrix(rnorm(100), nrow = 10, ncol = 10)
rownames(matrix) <- paste0("Gene", 1:10)
colnames(matrix) <- paste0("Sample", 1:10)
pHeatmap(matrix, scale = "row", dir_output = ".")
unlink("hm_matrix.pdf")
}
