% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_coord2}
\alias{plot_coord2}
\alias{plot_coord2.Seurat}
\title{plot_coord2}
\usage{
plot_coord2(x, ...)

\method{plot_coord2}{Seurat}(x, expand = NULL, ...)
}
\arguments{
\item{x}{object from which to plot coordinates.}

\item{...}{further arguments passed down to DimPlot().}

\item{expand}{logical; whether to expand one column to show presence/absence.}
}
\description{
Alternative version of plot_coord using Seurat::DimPlot internally for convenience.
Limited functionality at the moment, it only works with Seurat objects.
}
