% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpdb.R
\name{plot_cpdb_heatmap}
\alias{plot_cpdb_heatmap}
\title{plot_cpdb_heatmap}
\usage{
plot_cpdb_heatmap(x, cutoff = 0.001)
}
\arguments{
\item{cutoff}{cutoff for pvalues.}

\item{path}{path to the folder containing cellphonedb results.}
}
\description{
Plot cellphonedb results as a heatmap of cell types.
}
\details{
The score is calculated as the number of significant interactions (based on cutoff) per cell pair.
}
