% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diffusion_map.R
\name{plot_diffusion_map}
\alias{plot_diffusion_map}
\alias{add_dpt_paths}
\alias{plot_dpt_paths}
\title{Plot the diffusion map}
\usage{
plot_diffusion_map(
  x,
  colour_by = "DPT",
  add_paths = TRUE,
  root = NULL,
  paths_to = NULL,
  path_args = list(),
  w_width = 0.1,
  ...,
  ncomponents = 2L
)

add_dpt_paths(
  x,
  root = NULL,
  paths_to = NULL,
  w_width = 0.1,
  ...,
  ncomponents = 2L
)

plot_dpt_paths(...)
}
\arguments{
\item{x}{A \link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment}
object with \code{DiffusionMap} reducedDim component and \code{DPT} component if
add_paths is \code{TRUE}.}

\item{colour_by}{Specification of a column metadata field or a feature to
colour by, see the by argument in
\code{\link[scater:retrieveCellInfo]{?retrieveCellInfo}} for possible values. If
"DPT", \code{x} must contain components "destiny_branch" and "destiny_dpt", and
the DPT of root will be used.}

\item{add_paths}{A scalar logical value indicates whether to add path.}

\item{root}{Root branch ID. Will be used as the start of the DPT. (default:
lowest branch ID).}

\item{paths_to}{Numeric Branch IDs. Are used as target(s) for the path(s) to
draw.}

\item{path_args}{A list of arguments to passed to
\link[ggplot2:geom_path]{geom_path}.}

\item{w_width}{Window width for smoothing the path (see
\link[smoother:smth.gaussian]{smth.gaussian}).}

\item{...}{For \code{plot_diffusion_map}, Other arguments passed to
\link[scater:plotReducedDim]{plotReducedDim}; for \code{add_dpt_paths}, other
arguments passed to \link[ggplot2:geom_path]{geom_path}; and for
\code{plot_dpt_paths}, other arguments passed to \link{add_dpt_paths}}

\item{ncomponents}{A numeric scalar indicating the number of dimensions to
plot, starting from the first dimension. Alternatively, a numeric vector
specifying the dimensions to be plotted.}
}
\value{
\itemize{
\item \code{plot_diffusion_map}: A ggplot2 object depicting the Diffusion map, if
add_paths is \code{TRUE}, DPT of the root is also added in this map.
\item \code{plot_dpt_paths}: A ggplot2 object depicting the DPT paths.
\item \code{add_dpt_paths}: A ggplot2 geom object containing the DPT paths.
}
}
\description{
Plot the diffusion map
}
