% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{plot_enrichment}
\alias{plot_enrichment}
\title{plot_enrichment}
\usage{
plot_enrichment(
  x,
  group.by = "cluster",
  color.by = "P.Up",
  FDR = 0.01,
  col = c("white", "blue"),
  top_ann = NULL,
  ...
)
}
\arguments{
\item{x}{data.frame with results from run_enrichment.}

\item{group.by}{grouping variable (default: cluster).}

\item{color.by}{color variable (default: P.Up).}

\item{FDR}{cutoff to select significantly enriched terms/pathways.}

\item{col}{colors to use and passed down to ComplexHeatmap::Heatmap.}

\item{top_ann}{top annotations passed down to ComplexHeatmap::Heatmap.}

\item{...}{additional arguments passed down to ComplexHeatmap::Heatmap.}
}
\description{
Plot the result of run_enrichment function as a heatmap.
}
