% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_gene_modules}
\alias{plot_gene_modules}
\alias{plot_gene_modules.Seurat}
\title{plot_gene_modules}
\usage{
plot_gene_modules(x, ...)

\method{plot_gene_modules}{Seurat}(
  x,
  gene_modules,
  reduction = NULL,
  assay = NULL,
  slot = "data",
  size = 0.5,
  ...
)
}
\arguments{
\item{x}{object with gene expression levels per cell.}

\item{...}{arguments passed down to methods.}

\item{gene_modules}{data.frame with columns gene and module.}

\item{reduction}{reduction to use.}

\item{assay}{assay to use for Seurat objects.}

\item{slot}{slot to use for Seurat objects.}

\item{size}{size of the points.}
}
\description{
plot_gene_modules
}
