% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_gene_modules_heatmap}
\alias{plot_gene_modules_heatmap}
\alias{plot_gene_modules_heatmap.Seurat}
\title{plot_gene_modules_heatmap}
\usage{
plot_gene_modules_heatmap(x, ...)

\method{plot_gene_modules_heatmap}{Seurat}(
  x,
  gene_modules,
  group.by = "seurat_clusters",
  assay = NULL,
  slot = "data"
)
}
\arguments{
\item{x}{object with gene expression levels per cell.}

\item{...}{arguments passed down to methods.}

\item{gene_modules}{data.frame with columns gene and module.}

\item{group.by}{identity to use for grouping cells (default: seurat_clusters).}

\item{assay}{assay to use for Seurat objects.}

\item{slot}{slot to use for Seurat objects.}

\item{size}{size of the points.}
}
\description{
plot_gene_modules_heatmap
}
