% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_heat.R
\name{plot_grouped_heat}
\alias{plot_grouped_heat}
\alias{plot_grouped_heat,ANY-method}
\alias{plot_grouped_heat,SummarizedExperiment-method}
\title{Plot heatmap of group-level expression averages}
\usage{
plot_grouped_heat(x, ...)

\S4method{plot_grouped_heat}{ANY}(
  x,
  marker_list,
  clusters,
  cluster2cell = NULL,
  blocks = NULL,
  center = FALSE,
  scale = FALSE,
  zlim = NULL,
  colour = color,
  color = NULL,
  ...
)

\S4method{plot_grouped_heat}{SummarizedExperiment}(
  x,
  marker_list,
  ...,
  clusters = NULL,
  blocks = NULL,
  assay.type = "logcounts",
  swap_rownames = NULL
)
}
\arguments{
\item{x}{Currently, only
\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment} object is
supported.}

\item{...}{Other arguments passed to \link[ComplexHeatmap:Heatmap]{Heatmap}}

\item{marker_list}{A named list contaning the markers for each cell types.
names indicate the cell type label and values indicate the markers for this
cell type.}

\item{clusters}{A factor (or vector coercible into a factor) specifying the
group to which each cell in \code{x} belongs. Alternatively, if \code{x} is a
\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment}, e.g,
String specifying the field of \code{colData(x)} containing the grouping factor.
In this way, if clusters is \code{NULL}, "label" in \code{colData(x)} will be
extracted.}

\item{cluster2cell}{A named character or factor returned by
\code{\link[=annotate_clusters]{annotate_clusters()}}.}

\item{blocks}{A factor (or vector coercible into a factor) specifying the
blocking factor to which each cell in \code{x} belongs (e.g., batch of origin).
Alternatively, if \code{x} is a
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment}, e.g,
String specifying the field of \code{colData(x)} containing the grouping factor.
In this way, if clusters is \code{NULL}, "label" in \code{colData(x)} will be
extracted.}

\item{center}{A logical scalar indicating whether each feature should have its mean expression 
(specifically, the mean of averages across all groups) centered at zero prior to plotting.}

\item{scale}{A logical scalar specifying whether each row should have its
average expression values scaled to unit variance prior to plotting.}

\item{zlim}{A numeric vector of length 2, specifying the upper and lower bounds for color mapping of expression values.
Values outside this range are set to the most extreme color.
If \code{NULL}, it defaults to the range of the expression matrix.
If \code{center=TRUE}, this defaults to the range of the centered expression matrix, made symmetric around zero.}

\item{colour}{Alias for color.}

\item{color}{A vector of colours specifying the palette to use for increasing expression. 
This defaults to \link[viridis]{viridis} if \code{center=FALSE}, and the the \code{"RdYlBu"}
color palette from \code{\link[RColorBrewer]{brewer.pal}} otherwise.}

\item{assay.type}{A string or integer scalar indicating which \code{assays} in the
\code{x} contains the count matrix.}

\item{swap_rownames}{A characters (or vector coercible into a characters)
used to identify features instead of \code{rownames(x)} when labelling plot
elements.  Alternatively, if \code{x} is a
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment}, e.g,
String specifying the field of \code{rowData(x)} to be used.}
}
\description{
Create a heatmap of average expression values for each group of cells and
specified features in a SingleCellExperiment object.
}
