% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_heatmap}
\alias{plot_heatmap}
\alias{plot_heatmap.Seurat}
\alias{plot_heatmap.SingleCellExperiment}
\alias{plot_heatmap.matrix}
\title{plot_heatmap}
\usage{
plot_heatmap(x, ...)

\method{plot_heatmap}{Seurat}(
  x,
  assay = NULL,
  slot = "data",
  scale = TRUE,
  top_ann = NULL,
  top_ann_col = NULL,
  ...
)

\method{plot_heatmap}{SingleCellExperiment}(x, assay = "logcounts", top_ann = NULL, top_ann_col = NULL, ...)

\method{plot_heatmap}{matrix}(x, scale = TRUE, show_column_names = FALSE, ...)
}
\arguments{
\item{x}{object to plot.}

\item{...}{arguments passed down to ComplexHeatmap::Heatmap.}

\item{assay}{assay name for Seurat and SingleCellExperiment objects.}

\item{slot}{slot name for Seurat objects.}

\item{scale}{whether to scale data data.}

\item{top_ann}{names of columns to be used as top annotations.}

\item{top_ann_col}{color definition for the categories in the top annotations.}

\item{show_column_names}{whether to show column names (default: FALSE).}
}
\description{
plot_heatmap
}
