% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_jaccard}
\alias{plot_jaccard}
\alias{plot_jaccard.SingleCellExperiment}
\alias{plot_jaccard.DataFrame}
\alias{plot_jaccard.data.frame}
\title{plot_jaccard}
\usage{
plot_jaccard(x, ...)

\method{plot_jaccard}{SingleCellExperiment}(x, ...)

\method{plot_jaccard}{DataFrame}(x, ...)

\method{plot_jaccard}{data.frame}(x, col.x, col.y, label = FALSE, ...)
}
\arguments{
\item{x}{a suitable object.}

\item{...}{parameters passed down to methods.}

\item{col.x}{name of column to be use in x-axis.}

\item{col.y}{name of column to be use in y-axis.}
}
\description{
plot_jaccard
}
