% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_loadings}
\alias{plot_loadings}
\alias{plot_loadings.Seurat}
\title{plot_loadings}
\usage{
plot_loadings(x, features, reduction = "pca", cluster_columns = FALSE, ...)

\method{plot_loadings}{Seurat}(x, features, reduction = "pca", cluster_columns = FALSE, ...)
}
\arguments{
\item{x}{object to plot.}

\item{features}{name of features to plot.}

\item{reduction}{name of reduction to plot.}

\item{cluster_columns}{whether to cluster heatmap columns.}

\item{...}{arguments passed down to ComplexHeatmap::Heatmap.}
}
\description{
plot_loadings
}
