% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{plot_pseudotime_gene}
\alias{plot_pseudotime_gene}
\alias{plot_pseudotime_gene.cell_data_set}
\title{plot_pseudotime_gene}
\usage{
plot_pseudotime_gene(x, ...)

\method{plot_pseudotime_gene}{cell_data_set}(
  x,
  features = NULL,
  cutoff = 0,
  combine = TRUE,
  assay = "logcounts",
  ...
)
}
\arguments{
\item{x}{and object with pseudotime information.}

\item{...}{arguments passed down to methods.}
}
\description{
plot_pseudotime_gene
}
