% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{plot_pseudotime_heatmap}
\alias{plot_pseudotime_heatmap}
\alias{plot_pseudotime_heatmap.Seurat}
\title{plot_pseudotime_heatmap}
\usage{
plot_pseudotime_heatmap(x, ...)

\method{plot_pseudotime_heatmap}{Seurat}(
  x,
  features,
  assay = "RNA",
  slot = "data",
  reduction = "pseudotime",
  ...
)
}
\arguments{
\item{x}{and object with pseudotime information.}

\item{...}{arguments passed down to methods.}
}
\description{
plot_pseudotime_heatmap
}
