% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{plot_pseudotime_modules}
\alias{plot_pseudotime_modules}
\alias{plot_pseudotime_modules.Seurat}
\title{plot_pseudotime_modules}
\usage{
plot_pseudotime_modules(x, ...)

\method{plot_pseudotime_modules}{Seurat}(
  x,
  gene_modules,
  reduction = "pseudotime",
  assay = "RNA",
  slot = "data",
  filter.zero = TRUE,
  add.jitter = FALSE,
  width = 1
)
}
\arguments{
\item{x}{and object with pseudotime information.}

\item{...}{arguments passed down to methods.}
}
\description{
plot_pseudotime_modules
}
