% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haystack.R
\name{plot_result_rank}
\alias{plot_result_rank}
\alias{plot_result_rank.haystack}
\alias{plot_result_rank.data.frame}
\title{plot_result_rank}
\usage{
plot_result_rank(x, highlight = NULL, sort.by = "log.p.adj")

\method{plot_result_rank}{haystack}(x, highlight = NULL, sort.by = "log.p.adj")

\method{plot_result_rank}{data.frame}(x, highlight = NULL, sort.by = "log.p.adj")
}
\arguments{
\item{x}{haystack object or summary data.frame.}

\item{highlight}{gene names to highlight.}

\item{sort.by}{rank the genes by the following column from the summary data.frame.}
}
\description{
plot_result_rank
}
