% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{plot_spatial}
\alias{plot_spatial}
\alias{plot_spatial.Seurat}
\title{plot_spatial}
\usage{
plot_spatial(x, features, size = 0.5, rotate = FALSE)

\method{plot_spatial}{Seurat}(x, features, size = 0.5, rotate = FALSE)
}
\description{
plot_spatial
}
