% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{plot_trajectory}
\alias{plot_trajectory}
\alias{plot_trajectory.SingleCellExperiment}
\alias{plot_trajectory.SlingshotDataSet}
\title{plot_trajectory}
\usage{
plot_trajectory(x)

\method{plot_trajectory}{SingleCellExperiment}(x, ...)

\method{plot_trajectory}{SlingshotDataSet}(x)
}
\arguments{
\item{x}{and object with trajectory information.}
}
\description{
Plot the cell embedings with colors indicating the increasing trajectory pseudotime.
}
