% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{plot_trajectory_graph}
\alias{plot_trajectory_graph}
\alias{plot_trajectory_graph.SingleCellExperiment}
\alias{plot_trajectory_graph.SlingshotDataSet}
\alias{plot_trajectory_graph.tbl_graph}
\alias{plot_trajectory_graph.layout_ggraph}
\title{plot_trajectory_graph}
\usage{
plot_trajectory_graph(x, ...)

\method{plot_trajectory_graph}{SingleCellExperiment}(x, ...)

\method{plot_trajectory_graph}{SlingshotDataSet}(x, ...)

\method{plot_trajectory_graph}{tbl_graph}(x, layout = "nicely", ...)

\method{plot_trajectory_graph}{layout_ggraph}(x, ...)
}
\arguments{
\item{x}{and object with trajectory information.}

\item{...}{arguments passed down to methods.}

\item{layout}{layout for the graph.}
}
\description{
Plot the MST graph used to estimate trajectories.
}
