% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_velocity}
\alias{plot_velocity}
\title{plot_velocity}
\usage{
plot_velocity(
  x,
  meta = NULL,
  color = NULL,
  grid = FALSE,
  arrow.color = "black",
  arrow.size = 0.2
)
}
\arguments{
\item{x}{data.frame with velocity information.}

\item{meta}{cell information.}

\item{color}{column from meta to color cells.}

\item{grid}{whether to plot vector field.}

\item{arrow.color}{color for arrows.}

\item{arrow.size}{size for arrows.}
}
\description{
plot_velocity
}
