% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_violin}
\alias{plot_violin}
\alias{plot_violin.Seurat}
\alias{plot_violin.SingleCellExperiment}
\alias{plot_violin.data.frame}
\title{plot_violin}
\usage{
plot_violin(x, feature, group, ...)

\method{plot_violin}{Seurat}(x, feature, group, size = 0.1, ...)

\method{plot_violin}{SingleCellExperiment}(x, feature, group, size = 0.1,
  ...)

\method{plot_violin}{data.frame}(x, feature, group, size = 0.1, ...)
}
\arguments{
\item{x}{object to plot data from.}

\item{feature}{name of the feature to plot.}

\item{group}{grouping variable.}

\item{...}{further arguments passed down to get_expression().}

\item{size}{size of the points.}
}
\description{
Makes violin plot, by default using ggbeeswarm package.
}
