% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_volcano}
\alias{plot_volcano}
\title{plot_volcano}
\usage{
plot_volcano(x, clusters = NULL, n = 10, fdr = 0.01, lfc = 1, ...)
}
\arguments{
\item{x}{data.frame with DEG results from FindAllMarkers.}

\item{clusters}{clusters to plot. If null, plot all.}

\item{n}{number of genes (per down/up-regulated) to highlight.}

\item{fdr}{false discovery rate cutoff to highlight genes.}

\item{lfc}{log fold change cutoff to highlight genes.}

\item{...}{arguments passed down to ggrepel functions.}
}
\description{
plot_volcano
}
