% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDiffusionMap.R
\name{plotDiffusionMap}
\alias{plotDiffusionMap}
\title{Plot the diffusion map and the DPT paths}
\usage{
plotDiffusionMap(
  object,
  root = NULL,
  paths_to = NULL,
  add_paths = FALSE,
  path_args = list(),
  w_width = NULL,
  ...,
  dpt = NULL,
  colour_by = color_by,
  shape_by = NULL,
  size_by = NULL,
  order_by = NULL,
  text_by = NULL,
  color_by = NULL,
  dimred = "DiffusionMap",
  ncomponents = 2L
)
}
\arguments{
\item{object}{A
\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment} object.}

\item{root}{Root branch ID. Will be used as the start of the DPT. (default:
lowest branch ID)}

\item{paths_to}{Numeric Branch IDs, which are used as target(s) for the
path(s) to draw.}

\item{add_paths}{A scalar logical value indicates whether to add path.}

\item{path_args}{A list of arguments to passed to
\link[ggplot2:geom_path]{geom_path}.}

\item{w_width}{Window width for smoothing the path (see
\link[smoother:smth.gaussian]{smth.gaussian}). If \code{NULL}, will use the same
\code{w_width} used when running \link{calculateDiffusionMap}.}

\item{...}{
  Arguments passed on to \code{\link[scater:plotReducedDim]{scater::plotReducedDim}}
  \describe{
    \item{\code{by_exprs_values}}{Alias for \code{by.assay.type}.}
    \item{\code{text_size}}{Numeric scalar specifying the size of added text.}
    \item{\code{text_colour}}{String specifying the colour of the added text.}
    \item{\code{label_format}}{Character vector of length 2 containing format strings
to use for the axis labels.
The first string expects a string containing the result type (e.g.,
\code{"PCA"}) and an integer containing the component number,
while the second string shows the rounded percentage of variance explained
and is only relevant when this information is provided in \code{object}.}
    \item{\code{other_fields}}{Additional cell-based fields to include in the
data.frame, see \code{?"\link[scater]{scater-plot-args}"} for details.}
    \item{\code{text_color}}{Alias to \code{text_colour}.}
    \item{\code{swap_rownames}}{Column name of \code{rowData(object)} to be used to
identify features instead of \code{rownames(object)} when labelling plot
elements.}
    \item{\code{point.padding,force}}{See \code{?ggrepel::geom_text_repel}.}
    \item{\code{rasterise}}{Whether to rasterise the points in the plot with
\code{\link[ggrastr]{rasterise}}. To control the dpi, set
\code{options(ggrastr.default.dpi)},
for example \code{options(ggrastr.default.dpi=300)}.}
    \item{\code{scattermore}}{Logical, whether to use the \code{scattermore} package to
greatly speed up plotting a large number of cells. Use \code{point_size =
0} for the most performance gain.}
    \item{\code{bins}}{Number of bins, can be different in x and y, to bin and summarize
the points and their values, to avoid overplotting. If \code{NULL}
(default), then the points are plotted without binning. Only used when both
x and y are numeric.}
    \item{\code{summary_fun}}{Function to summarize the feature value of each point
(e.g. gene expression of each cell) when the points binned, defaults to
\code{sum}. Can be either the name of the function or the function itself.}
    \item{\code{hex}}{Logical, whether to use \code{\link{geom_hex}}. Note that
\code{geom_hex} is broken in \code{ggplot2} version 3.4.0.}
    \item{\code{by.assay.type}}{A string or integer scalar specifying which assay to
obtain expression values from,
for use in point aesthetics - see the \code{assay.type} argument in
\code{?\link[scater]{retrieveCellInfo}}.}
  }}

\item{dpt}{A string, specify the \code{DPT} to use, usually starts with "DPT". If
\code{NULL}, will use \code{root} to extract the \code{DPT}. \code{colour_by}, \code{shape_by},
\code{size_by}, \code{order_by}, \code{text_by} and \code{color_by} can use "DPT" to mapping from
the extracted \code{DPT}.}

\item{colour_by}{Specification of a column metadata field or a feature to
colour by, see the \code{by} argument in \code{?\link[scater]{retrieveCellInfo}} for
possible values.}

\item{shape_by}{Specification of a column metadata field or a feature to
shape by, see the \code{by} argument in \code{?\link[scater]{retrieveCellInfo}} for
possible values.}

\item{size_by}{Specification of a column metadata field or a feature to
size by, see the \code{by} argument in \code{?\link[scater]{retrieveCellInfo}} for
possible values.}

\item{order_by}{Specification of a column metadata field or a feature to
order points by, see the \code{by} argument in
\code{?\link[scater]{retrieveCellInfo}} for possible values.}

\item{text_by}{String specifying the column metadata field with which to add
text labels on the plot.
This must refer to a categorical field, i.e., coercible into a factor.
Alternatively, an \link{AsIs} vector or data.frame, see
\code{?\link[scater]{retrieveCellInfo}}.}

\item{color_by}{Alias to \code{colour_by}.}

\item{dimred}{A string or integer scalar indicating the reduced dimension
result in \code{reducedDims(object)} to plot.}

\item{ncomponents}{A numeric scalar indicating the number of dimensions to
plot, starting from the first dimension.
Alternatively, a numeric vector specifying the dimensions to be plotted.}
}
\value{
A ggplot object
}
\description{
Plot the diffusion map and the DPT paths
}
