% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotSlingshot}
\alias{plotSlingshot}
\title{plot slingshot results}
\usage{
plotSlingshot(object, lineage)
}
\arguments{
\item{object}{A Seurat object}

\item{lineage}{A character string indicating the lineage of interest.}
}
\value{
A ggplot object.
}
\description{
The function creates a colored UMAP plot and curves split by lineage
}
\examples{
\dontrun{
sds_plots_list <- lapply(colnames(pt), slingshotPlot, object = bcells)
}
}
