% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_fs.R
\name{pool_fs}
\alias{pool_fs}
\title{Pool a flowSet of multiple flowFrames into a single pooled flowFrame.}
\usage{
pool_fs(fs, copy_desc = TRUE)
}
\arguments{
\item{fs}{Input flowSet.}

\item{copy_desc}{(Optional): if `TRUE` will attempt to copy common keywords
to output `flowFrame`.}
}
\value{
A pooled flowFrame.
}
\description{
`pool_fs()` lets you pool cells from multiple flowFrames contained in
a `flowSet` into a single `flowFrame`. If `copy_desc = TRUE` it also
attempts to copy the keyword values that are common in all constituent
flowFrames.
}
\note{
The input fs can either be a `flowCore::flowSet` or `ncdfFlow::ncdfFlowSet`.
}
\examples{
# Mock up two flowFrames
set.seed(1234)
library(flowCore)
rstrings <- function(n = 10) {
  a <- do.call(paste0, replicate(5, sample(LETTERS, n, TRUE), FALSE))
  paste0(a, sprintf("\%04d", sample(9999, n, TRUE)), sample(LETTERS, n, TRUE))
}
ff1_exprs <- matrix(rnorm(1:10000), nrow = 1000)
colnames(ff1_exprs) <- rstrings(dim(ff1_exprs)[[2]])
ff1 <- flowFrame(ff1_exprs)
ff2_exprs <- matrix(rnorm(1:10000), nrow = 1000)
colnames(ff2_exprs) <- colnames(ff1_exprs)
ff2 <- flowFrame(ff2_exprs)
fs <- flowSet(ff1, ff2)
dim(ff1)
#>     events parameters
#>       1000         10
dim(ff2)
#>     events parameters
#>       1000         10
fs <- flowSet(ff1, ff2)
pooled_ff <- pool_fs(fs)
dim(pooled_ff)
#>     events parameters
#>       2000         10

}
