% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process}
\alias{process}
\alias{process.Seurat}
\title{process}
\usage{
process(x, ...)

\method{process}{Seurat}(
  x,
  assay = NULL,
  dims = 1:10,
  algorithm = 1,
  resolution = 0.6,
  verbose = FALSE
)
}
\arguments{
\item{x}{an object of class Seurat.}

\item{assay}{assay to use for processing.}

\item{dims}{PCA dimensions to use for UMAP and clustering.}

\item{algorithm}{algorithm to use for clustering.}

\item{resolution}{resolution to use for clustering.}

\item{verbose}{whether to output diagnostic information.}
}
\description{
Applies several processing steps to a single cell genomics object.
}
