% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_helper.R
\name{propellerCalc}
\alias{propellerCalc}
\title{propellerCalc}
\usage{
propellerCalc(
  seu_obj1,
  condition1,
  condition2,
  cluster_col,
  meta_col,
  lookup,
  sample_col,
  formula,
  min_cells = 30,
  adjustment_method = "BH",
  fdr_threshold = NULL,
  n_perms = 1000
)
}
\arguments{
\item{seu_obj1}{A seurat object}

\item{condition1}{A character representing the first condition}

\item{condition2}{A character representing the second condition}

\item{cluster_col}{A character representing the name of the column which contains cluster information}

\item{meta_col}{A character representing the name of the column which contains meta information, should be in Seurat object and in lookup table}

\item{lookup}{A dataframe that contains sample information}

\item{sample_col}{A character representing the name of the column in seu_obj1 which contains sample information}

\item{formula}{A linear model that should be used for the design matrix}

\item{min_cells}{A numeric value indicating the minimum number of cells in a cluster that should be included in the analysis}

\item{adjustment_method}{A character indicating the FDR adjustment method: "BH" (default) or "permFDP"}

\item{fdr_threshold}{The FDR threshold to use for permFDP adjustment (default: NULL)}

\item{n_perms}{Number of permutations for permFDP adjustment (default: 1000)}
}
\value{
A dataframe containing the results of the t-test
}
\description{
perform a propeller t-test on a Seurat object
}
\examples{
set.seed(123)
library(Seurat)
pbmc_small$condition <- factor(sample(c("diseaseA", "diseaseB"), nrow(pbmc_small), replace = TRUE))
pbmc_small$cluster <- Idents(pbmc_small)
pbmc_small$patient <- rep(paste0("P", 0:9), each = 8, length.out = ncol(pbmc_small))
lookup <- data.frame(
  patient = paste0("P", 0:9),
  condition = sample(c("diseaseA", "diseaseB"), 10, replace = TRUE)
)
# Standard BH adjustment
propellerCalc(
  seu_obj1 = pbmc_small,
  condition1 = "diseaseA",
  condition2 = "diseaseB",
  cluster_col = "cluster",
  meta_col = "condition",
  lookup = lookup,
  sample_col = "patient",
  formula = "~ 0 + condition",
  min_cells = 30
)

# With permFDP adjustment
propellerCalc(
  seu_obj1 = pbmc_small,
  condition1 = "diseaseA",
  condition2 = "diseaseB",
  cluster_col = "cluster",
  meta_col = "condition",
  lookup = lookup,
  sample_col = "patient",
  formula = "~ 0 + condition",
  min_cells = 30,
  adjustment_method = "permFDP",
  fdr_threshold = 0.1,
  n_perms = 100
)
}
