% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobulk.R
\name{pseudobulk}
\alias{pseudobulk}
\alias{pseudobulk.Seurat}
\title{pseudobulk}
\usage{
pseudobulk(x, split.by, group.by, samples = NULL, genes = NULL, ...)

\method{pseudobulk}{Seurat}(
  x,
  split.by,
  group.by,
  samples = NULL,
  genes = NULL,
  assay = NULL,
  layers = "counts",
  ...
)
}
\arguments{
\item{x}{Seurat object.}

\item{split.by}{Split single cell data using this column. Typically cell populations or clusters.}

\item{group.by}{Aggregate counts from cells in each group. Typically sample names.}

\item{samples}{A data.frame with sample information to include in slot "samples" in the DGEList object.}

\item{genes}{A data.frame with gene information to include in slot "genes" in the DGEList object.}

\item{assay}{Seurat assay to use. Default: NULL (default assay).}

\item{layers}{Seurat layers to use. Default: counts.}
}
\description{
Generate pseudobulk counts of single cell data per cell population and sample name. Return a list of DGEList objects.
}
