% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpdb.R
\name{read_cpdb_out}
\alias{read_cpdb_out}
\title{read_cpdb_out}
\usage{
read_cpdb_out(path)
}
\arguments{
\item{path}{path to the folder containing cellphonedb results.}
}
\description{
Read cellphonedb results.
}
