% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_helper.R
\name{ReadCellBender_h5}
\alias{ReadCellBender_h5}
\title{Load CellBender h5 matrices}
\usage{
ReadCellBender_h5(file_name, use.names = TRUE, unique.features = TRUE)
}
\arguments{
\item{file_name}{Path to h5 file.}

\item{use.names}{Label row names with feature names rather than ID numbers (default TRUE).}

\item{unique.features}{Make feature names unique (default TRUE).}
}
\value{
sparse matrix
}
\description{
Extract sparse matrix with corrected counts from CellBender h5 output file.
}
\examples{
file_name <- system.file("extdata", "raw_feature_bc_matrix_filtered.h5", package = "scMisc")
mat <- ReadCellBender_h5(file_name)
}
\references{
Code used in function has been modified from \code{Seurat::Read10X_h5} function of
Seurat package \url{https://github.com/satijalab/seurat} (License: GPL-3).
}
