% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DotsHeatmap.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{gpar}
\alias{draw}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ComplexHeatmap}{\code{\link[ComplexHeatmap:draw-dispatch]{draw}}}

  \item{grid}{\code{\link[grid]{gpar}}}
}}

