% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_diffusion_map.R
\name{run_diffusion_map}
\alias{run_diffusion_map}
\alias{run_diffusion_map,ANY-method}
\alias{run_diffusion_map,SingleCellExperiment-method}
\title{Create a diffusion map of cells}
\usage{
run_diffusion_map(x, ...)

\S4method{run_diffusion_map}{ANY}(x, dpt = TRUE, ...)

\S4method{run_diffusion_map}{SingleCellExperiment}(
  x,
  dimred = "PCA",
  ...,
  dpt = TRUE,
  n_dimred = NULL,
  assay.type = "logcounts"
)
}
\arguments{
\item{x}{A numeric matrix of counts with cells in columns and features in
rows.

Alternatively, a
\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment} object
containing such a matrix.}

\item{...}{Other arguments passed to \link[destiny:DiffusionMap-class]{destiny}}

\item{dpt}{A scalar logical value indicates whether to run
\link[destiny:DPT]{DPT}.}

\item{dimred}{String or integer scalar specifying the existing dimensionality
reduction results to use.}

\item{n_dimred}{Integer scalar or vector specifying the dimensions to use if
dimred is specified.}

\item{assay.type}{A string or integer scalar indicating which \code{assays} in the
\code{x} contains the count matrix. Only used when dimred is \code{NULL}.}
}
\value{
A list or a modified x that contains the DiffusionMap coordinates in
reducedDim(x, "DiffusionMap"), if dpt is TRUE, branch and dpt are also kept
in \code{colData(x)}.
}
\description{
A modified version of \link[destiny:DiffusionMap-class]{DiffusionMap}, this function
will add the DiffusionMap results components into x if it is a
\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment} object.
}
