% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{run_enrichment}
\alias{run_enrichment}
\alias{run_enrichment.data.frame}
\title{run_enrichment}
\usage{
run_enrichment(x, ...)

\method{run_enrichment}{data.frame}(
  x,
  type = "kegg",
  group.by = "cluster",
  use.column = "entrezgene",
  org = "Mm",
  lfc = 1,
  fdr = 0.01,
  p.adjust.method = "bonferroni",
  ...
)
}
\arguments{
\item{x}{an object with DEGs.}

\item{...}{arguments passed down to methods.}

\item{type}{which ontology to use for enrichment (default: KEGG).}

\item{group.by}{which variable to use for grouping (default: cluster).}

\item{use.column}{which column contains entrez gene identifiers (default: entrezgene).}

\item{org}{organisms (default: Mm).}

\item{lfc}{avg_log2FC cutoff for selection of DEG used for enrichment.}

\item{fdr}{p_val_adj cutoff for selection of DEG used for enrichment.}

\item{p.adjust.method}{method to adjust p.values from enrichment (default: bonferroni).}
}
\description{
run_enrichment
}
