% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infercnv.R
\name{run_infercnv}
\alias{run_infercnv}
\title{Run infercnv workflow}
\usage{
run_infercnv(
  counts,
  granges,
  groups,
  ref_groups = character(),
  keytype = NULL,
  standard_chr = TRUE,
  species = NULL,
  seqstyle = "UCSC",
  chr_exclude = c("chrX", "chrY", "chrM"),
  max_cells_per_group = NULL,
  min_max_counts_per_cell = c(100, Inf),
  png_res = 600L,
  ...
)
}
\arguments{
\item{counts}{Single cell gene expression counts matrix.}

\item{granges}{A \link[GenomicRanges:GRanges-class]{GRanges} object indicates the
positions of each gene. It's also possible to provide a
\link[ensembldb:EnsDb]{EnsDb} or \link[GenomicFeatures:TxDb-class]{TxDb} object, in
this way, \code{keytype} will be used to define the gene identifiers.}

\item{groups}{A character indicates the cell type of each column.}

\item{ref_groups}{Groups used as the reference.}

\item{keytype}{The gene type in the counts matrix, either "gene_id" or
"symbol". This parameter is only applicable when the \code{granges} object is an
\link[ensembldb:EnsDb]{EnsDb} with the default value set to "symbol", or a
\link[GenomicFeatures:TxDb-class]{TxDb} with the default value set to "gene_id".
When "gene_id" is used, it represents the ensemble id for
\link[ensembldb:EnsDb]{EnsDb} or entrez id for
\link[GenomicFeatures:TxDb-class]{TxDb}.}

\item{standard_chr}{A boolean value indicates whether keep only 'standard'
chromosomes. See
\link[GenomeInfoDb:seqlevels-wrappers]{keepStandardChromosomes}.}

\item{species}{A string indicates the genus and species of the organism.
Supported species can be seen with \code{names(GenomeInfoDb::genomeStyles())}.
Only used when standard_chr is \code{TRUE}.}

\item{seqstyle}{A character string that sets the seqlevels style for
\code{granges}.}

\item{chr_exclude}{list of chromosomes in the reference genome annotations that should be excluded from analysis.  Default = c('chrX', 'chrY', 'chrM')}

\item{max_cells_per_group}{maximun number of cells to use per group. Default=NULL, using all cells defined in the annotations_file. This option is useful for randomly subsetting the existing data for a quicker preview run, such as using 50 cells per group instead of hundreds.}

\item{min_max_counts_per_cell}{minimum and maximum counts allowed per cell. Any cells outside this range will be removed from the counts matrix. default=(100, +Inf) and uses all cells. If used, should be set as c(min_counts, max_counts)}

\item{png_res}{Resolution for png output.}

\item{...}{
  Arguments passed on to \code{\link[infercnv:run]{infercnv::run}}
  \describe{
    \item{\code{cutoff}}{Cut-off for the min average read counts per gene among reference cells. (default: 1)}
    \item{\code{min_cells_per_gene}}{minimum number of reference cells requiring expression measurements to include the corresponding gene.
default: 3}
    \item{\code{out_dir}}{path to directory to deposit outputs (default: NULL, required to provide non NULL)

## Smoothing params}
    \item{\code{window_length}}{Length of the window for the moving average
(smoothing). Should be an odd integer. (default: 101)#'}
    \item{\code{smooth_method}}{Method to use for smoothing: c(runmeans,pyramidinal,coordinates)  default: pyramidinal

#####}
    \item{\code{num_ref_groups}}{The number of reference groups or a list of
indices for each group of reference indices in
relation to reference_obs. (default: NULL)}
    \item{\code{ref_subtract_use_mean_bounds}}{Determine means separately for each ref group, then remove intensities within bounds of means (default: TRUE)
                                      Otherwise, uses mean of the means across groups.

#############################}
    \item{\code{cluster_by_groups}}{If observations are defined according to groups (ie. patients), each group
of cells will be clustered separately. (default=FALSE, instead will use k_obs_groups setting)}
    \item{\code{cluster_references}}{Whether to cluster references within their annotations or not. (dendrogram not displayed)
(default: TRUE)}
    \item{\code{k_obs_groups}}{Number of groups in which to break the observations. (default: 1)}
    \item{\code{hclust_method}}{Method used for hierarchical clustering of cells. Valid choices are:
"ward.D", "ward.D2", "single", "complete", "average", "mcquitty", "median", "centroid".
default("ward.D2")}
    \item{\code{max_centered_threshold}}{The maximum value a value can have after
    centering. Also sets a lower bound of
    -1 * this value. (default: 3),
can set to a numeric value or "auto" to bound by the mean bounds across cells.
Set to NA to turn off.}
    \item{\code{scale_data}}{perform Z-scaling of logtransformed data (default: FALSE).  This may be turned on if you have
                   very different kinds of data for your normal and tumor samples. For example, you need to use GTEx
                   representative normal expression profiles rather than being able to leverage normal single cell data
                   that goes with your experiment.

#########################################################################
## Downstream Analyses (HMM or non-DE-masking) based on tumor subclusters}
    \item{\code{HMM}}{when set to True, runs HMM to predict CNV level (default: FALSE)}
    \item{\code{HMM_transition_prob}}{transition probability in HMM (default: 1e-6)}
    \item{\code{HMM_report_by}}{cell, consensus, subcluster (default: subcluster)  Note, reporting is performed entirely separately from the HMM prediction.  So, you can predict on subclusters, but get per-cell level reporting (more voluminous output).}
    \item{\code{HMM_type}}{HMM model type. Options: (i6 or i3):
i6: infercnv 6-state model (0, 0.5, 1, 1.5, 2, >2) where state emissions are calibrated based on simulated CNV levels.
i3: infercnv 3-state model (del, neutral, amp) configured based on normal cells and HMM_i3_pval}
    \item{\code{HMM_i3_pval}}{p-value for HMM i3 state overlap (default: 0.05)}
    \item{\code{HMM_i3_use_KS}}{boolean: use the KS test statistic to estimate mean of amp/del distributions (ala HoneyBadger). (default=TRUE)


## Filtering low-conf HMM preds via BayesNet P(Normal)}
    \item{\code{BayesMaxPNormal}}{maximum P(Normal) allowed for a CNV prediction according to BayesNet. (default=0.5, note zero turns it off)}
    \item{\code{sim_method}}{method for calibrating CNV levels in the i6 HMM (default: 'meanvar')}
    \item{\code{sim_foreground}}{don't use... for debugging, developer option.}
    \item{\code{reassignCNVs}}{(boolean) Given the CNV associated probability of belonging to each possible state, 
                     reassign the state assignments made by the HMM to the state that has the highest probability. (default: TRUE)

######################
## Tumor subclustering}
    \item{\code{analysis_mode}}{options(samples|subclusters|cells), Grouping level for image filtering or HMM predictions.
default: samples (fastest, but subclusters is ideal)}
    \item{\code{tumor_subcluster_partition_method}}{method for defining tumor subclusters. Options('leiden', 'random_trees', 'qnorm')
leiden: Runs a nearest neighbor search, where communities are then partitionned with the Leiden algorithm.
random_trees: Slow, uses permutation statistics w/ tree construction.
qnorm: defines tree height based on the quantile defined by the tumor_subcluster_pval}
    \item{\code{tumor_subcluster_pval}}{max p-value for defining a significant tumor subcluster (default: 0.1)}
    \item{\code{k_nn}}{number k of nearest neighbors to search for when using the Leiden partition method for subclustering (default: 20)}
    \item{\code{leiden_method}}{Method used to generate the graph on which the Leiden algorithm is applied, one of "PCA" or "simple". (default: "PCA")}
    \item{\code{leiden_function}}{Whether to use the Constant Potts Model (CPM) or modularity in igraph. Must be either "CPM" or "modularity". (default: "CPM")}
    \item{\code{leiden_resolution}}{resolution parameter for the Leiden algorithm using the CPM quality score (default: auto)}
    \item{\code{leiden_method_per_chr}}{Method used to generate the graph on which the Leiden algorithm is applied for the per chromosome subclustering, one of "PCA" or "simple". (default: "simple")}
    \item{\code{leiden_function_per_chr}}{Whether to use the Constant Potts Model (CPM) or modularity in igraph for the per chromosome subclustering. Must be either "CPM" or "modularity". (default: "modularity")}
    \item{\code{leiden_resolution_per_chr}}{resolution parameter for the Leiden algorithm for the per chromosome subclustering (default: 1)}
    \item{\code{per_chr_hmm_subclusters}}{Run subclustering per chromosome over all cells combined to run the HMM on those subclusters instead. Only applicable when using Leiden subclustering.
This should provide enough definition in the predictions while avoiding subclusters that are too small thus providing less evidence to work with. (default: FALSE)}
    \item{\code{per_chr_hmm_subclusters_references}}{Whether the per chromosome subclustering should also be done on references, which should not have as much variation as observations. (default = FALSE)}
    \item{\code{z_score_filter}}{Z-score used as a treshold to filter genes used for subclustering. 
                      Applied based on reference genes to automatically ignore genes with high expression variability such as MHC genes. (default: 0.8)


#############################
## de-noising parameters ####}
    \item{\code{denoise}}{If True, turns on denoising according to options below}
    \item{\code{noise_filter}}{Values +- from the reference cell mean will be set to zero (whitening effect)
default(NA, instead will use sd_amplifier below.}
    \item{\code{sd_amplifier}}{Noise is defined as mean(reference_cells) +- sdev(reference_cells) * sd_amplifier
default: 1.5}
    \item{\code{noise_logistic}}{use the noise_filter or sd_amplifier based threshold (whichever is invoked) as the midpoint in a
                      logistic model for downscaling values close to the mean. (default: FALSE)


##################
## Outlier pruning}
    \item{\code{outlier_method_bound}}{Method to use for bounding outlier values. (default: "average_bound")
Will preferentially use outlier_lower_bounda and outlier_upper_bound if set.}
    \item{\code{outlier_lower_bound}}{Outliers below this lower bound will be set to this value.}
    \item{\code{outlier_upper_bound}}{Outliers above this upper bound will be set to this value.


##########################
## Misc options}
    \item{\code{final_scale_limits}}{The scale limits for the final heatmap output by the run() method. Default "auto". Alt, c(low,high)}
    \item{\code{final_center_val}}{Center value for final heatmap output by the run() method.}
    \item{\code{debug}}{If true, output debug level logging.}
    \item{\code{num_threads}}{(int) number of threads for parallel steps (default: 4)}
    \item{\code{plot_steps}}{If true, saves infercnv objects and plots data at the intermediate steps.}
    \item{\code{inspect_subclusters}}{If true, plot subclusters as annotations after the subclustering step to easily see if the subclustering options are good. (default = TRUE)}
    \item{\code{resume_mode}}{leverage pre-computed and stored infercnv objects where possible. (default=TRUE)}
    \item{\code{plot_probabilities}}{option to plot posterior probabilities (default: TRUE)}
    \item{\code{save_rds}}{Whether to save the current step object results as an .rds file (default: TRUE)}
    \item{\code{save_final_rds}}{Whether to save the final object results as an .rds file (default: TRUE)}
    \item{\code{diagnostics}}{option to create diagnostic plots after running the Bayesian model (default: FALSE)

#######################
## Experimental options}
    \item{\code{remove_genes_at_chr_ends}}{experimental option: If true, removes the window_length/2 genes at both ends of the chromosome.}
    \item{\code{prune_outliers}}{Define outliers loosely as those that exceed the mean boundaries among all cells.  These are set to the bounds.

## experimental opts involving DE analysis}
    \item{\code{mask_nonDE_genes}}{If true, sets genes not significantly differentially expressed between tumor/normal to
the mean value for the complete data set (default: 0.05)}
    \item{\code{mask_nonDE_pval}}{p-value threshold for defining statistically significant DE genes between tumor/normal}
    \item{\code{test.use}}{statistical test to use.  (default: "wilcoxon") alternatives include 'perm' or 't'.'}
    \item{\code{require_DE_all_normals}}{If mask_nonDE_genes is set, those genes will be masked only if they are are found as DE according to test.use and mask_nonDE_pval in each of the comparisons to normal cells options: {"any", "most", "all"} (default: "any")

other experimental opts}
    \item{\code{hspike_aggregate_normals}}{instead of trying to model the different normal groupings individually, just merge them in the hspike.}
    \item{\code{no_plot}}{don't make any of the images. Instead, generate all non-image outputs as part of the run. (default: FALSE)}
    \item{\code{no_prelim_plot}}{don't make the preliminary infercnv image (default: FALSE)}
    \item{\code{write_expr_matrix}}{Whether to write text files with the content of matrices when generating plots (default: FALSE)}
    \item{\code{write_phylo}}{Whether to write newick strings of the dendrograms displayed on the left side of the heatmap to file (default: FALSE)}
    \item{\code{output_format}}{Output format for the figure. Choose between "png", "pdf" and NA. NA means to only write the text outputs without generating the figure itself. (default: "png")}
    \item{\code{plot_chr_scale}}{Whether to scale the chromosme width on the heatmap based on their actual size rather than just the number of expressed genes.}
    \item{\code{chr_lengths}}{A named list of chromsomes lengths to use when plot_chr_scale=TRUE, or else chromosome size is assumed to be the last chromosome's stop position + 10k bp}
    \item{\code{useRaster}}{Whether to use rasterization for drawing heatmap. Only disable if it produces an error as it is much faster than not using it. (default: TRUE)}
    \item{\code{up_to_step}}{run() only up to this exact step number (default: 100 >> 23 steps currently in the process)}
  }}
}
\description{
Run infercnv workflow
}
\references{
\url{https://github.com/broadinstitute/infercnv/wiki}
}
\seealso{
\itemize{
\item \code{\link[infercnv:CreateInfercnvObject]{infercnv::CreateInfercnvObject()}}
\item \code{\link[infercnv:run]{infercnv::run()}}
}
}
