% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample_cells}
\alias{sample_cells}
\alias{sample_cells.Seurat}
\alias{sample_cells.SingleCellExperiment}
\title{sample_cells}
\usage{
sample_cells(x, ...)

\method{sample_cells}{Seurat}(x, group = NULL, n = NULL, frac = NULL, n_max = NULL, ...)

\method{sample_cells}{SingleCellExperiment}(x, group = NULL, n = NULL, frac = NULL, ...)
}
\arguments{
\item{x}{an object to sample from.}

\item{...}{parameters passed down to methods.}

\item{group}{optional grouping variable.}

\item{n}{number of cells to sample (per group).}

\item{frac}{fraction of cells to sample (per group).}

\item{n_max}{number of cells to sample, if less cells available then all cells returned.}
}
\description{
Subset a single cell experiment object by sampling the number of specified cells,
optionally by grouping.
}
