% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helper.R
\name{scale_mat}
\alias{scale_mat}
\title{scale matrix from pheatmap only internal use}
\usage{
scale_mat(mat, scale)
}
\arguments{
\item{mat}{matrix input}

\item{scale}{should the values be centered and scaled in row, column direction? Allowed: \code{row}, \code{column}, \code{none}}
}
\value{
matrix scaled
}
\description{
copied from pheatmap for internal use
}
\examples{
set.seed(123)
mat <- matrix(rnorm(100), 10, 10)
scale_mat(mat, "row")
}
