% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helper.R
\name{scale_rows}
\alias{scale_rows}
\title{scale rows from pheatmap only internal use}
\usage{
scale_rows(x)
}
\arguments{
\item{x}{matrix input}
}
\value{
matrix scaled rows
}
\description{
copied from pheatmap for internal use
}
\examples{
set.seed(123)
x <- matrix(rnorm(100), 10, 10)
scale_rows(x)
}
