% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-palettes.R
\name{scale_color_scanpy}
\alias{scale_color_scanpy}
\alias{scale_colour_scanpy}
\alias{scale_fill_scanpy}
\title{SCANPY Color Scales}
\usage{
scale_color_scanpy(palette = c("zeileis", "godsnot"), alpha = 1, ...)

scale_colour_scanpy(palette = c("zeileis", "godsnot"), alpha = 1, ...)

scale_fill_scanpy(palette = c("zeileis", "godsnot"), alpha = 1, ...)
}
\arguments{
\item{palette}{Palette type.
Currently there is one available option: \code{"default"} (102-color palette).}

\item{alpha}{Transparency level, a real number in \verb{(0, 1]}.
See \code{alpha} in \link[grDevices:rgb]{rgb} for details.}

\item{...}{additional parameters for
\link[ggplot2:discrete_scale]{discrete_scale}}
}
\description{
See \code{\link{pal_scanpy}} for details.
}
