% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmisc-package.R
\docType{package}
\name{scmisc-package}
\alias{scmisc}
\alias{scmisc-package}
\title{scmisc: Miscellaneous functions for Single Cell Analysis}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Yun Peng \email{yunyunpp96@outlook.com}

}
\keyword{internal}
