% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scprop.R
\name{scprop_compare}
\alias{scprop_compare}
\title{Permutation Test For Proportions}
\usage{
scprop_compare(
  identity,
  compare,
  stat = "risk_ratio",
  times = 2000L,
  n_permutation = times,
  n_bootstrap = times,
  p.adjust = "BH",
  conf.int = TRUE,
  conf.level = 0.95,
  BPPARAM = BiocParallel::SerialParam()
)
}
\arguments{
\item{identity}{An atomic vector define the cell types}

\item{compare}{An atomic vector coerced into factor to define the comparison
groups, must have two unique values.}

\item{stat}{A string giving the type of the statistic to calculate. Current
options include "odds_ratio", "risk_ratio", "risk_diff".}

\item{times}{Conventient way to set both \code{n_permutation} and \code{n_bootstrap}.}

\item{n_permutation}{Number of permutation to calcualte \code{p.value}.}

\item{n_bootstrap}{Number of bootstrap to calcualte \code{ci.low} and \code{ci.high}.}

\item{p.adjust}{Correction method, a string. Can be abbreviated. See
\link[stats:p.adjust]{p.adjust}.}

\item{conf.int}{Logical indicating whether or not to include a confidence
interval in the tidied output. Defaults to \code{TRUE}.}

\item{conf.level}{The confidence level to use for the confidence interval if
conf.int = \code{TRUE}. Must be strictly greater than 0 and less than 1. Defaults
to 0.95, which corresponds to a 95 percent confidence interval.}

\item{BPPARAM}{A \link[BiocParallel:BiocParallelParam-class]{BiocParallelParam} object
specifying whether Permutation (and Bootstraping) should be parallelized.}
}
\value{
A \link[data.table:data.table]{data.table}
}
\description{
Permutation Test For Proportions
}
\seealso{
\url{https://github.com/rpolicastro/scProportionTest}
}
