% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scprop_edger.R
\name{scprop_edger}
\alias{scprop_edger}
\alias{scprop_edger,data.frame-method}
\alias{scprop_edger,SummarizedExperiment-method}
\title{Changes in cluster abundance}
\usage{
scprop_edger(data, ...)

\S4method{scprop_edger}{data.frame}(
  data,
  formula,
  sample,
  celltype,
  filter = TRUE,
  normalize = FALSE,
  coef = NULL,
  p.adjust = "BH"
)

\S4method{scprop_edger}{SummarizedExperiment}(data, ..., sample = "Sample", celltype = "label")
}
\arguments{
\item{data}{A data.frame.}

\item{...}{For the generic, further arguments to pass to specific methods.}

\item{formula}{A formula object to build design matrix from \code{data}.}

\item{sample}{A string of the column in \code{data} containing sample id.}

\item{celltype}{A string of the column in \code{data} containing cell types.}

\item{filter}{A bool, if \code{TRUE}, will use \link[edgeR:filterByExpr]{filterByExpr}
to filter.}

\item{normalize}{A bool, if \code{TRUE}, will use
\link[edgeR:calcNormFactors]{calcNormFactors} to compute normalization factors.}

\item{coef}{An integer or character index vector indicating which
coefficients of the linear model are to be tested equal to zero. If \code{NULL},
the last coef will be used.}

\item{p.adjust}{Correction method, a string. Can be abbreviated. See
\link[stats:p.adjust]{p.adjust}.}
}
\value{
A \link[data.table:data.table]{data.table}.
}
\description{
Changes in cluster abundance
}
\references{
\url{https://bioconductor.org/books/3.18/OSCA.multisample/differential-abundance.html#}
}
