% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DotsHeatmap.R
\name{show,DotsHeatmap-method}
\alias{show,DotsHeatmap-method}
\title{Draw a DotsHeatmap with default parameters}
\usage{
\S4method{show}{DotsHeatmap}(object)
}
\arguments{
\item{object}{A \link{DotsHeatmap} object.}
}
\description{
Draw a DotsHeatmap with default parameters
}
\details{
Actually it calls \link{draw}, but only with default parameters. If users
want to customize the DotsHeatmap, they can pass parameters directly to
\link{draw}.
}
