% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{stackedPlot}
\alias{stackedPlot}
\title{abundance stacked bar plot}
\usage{
stackedPlot(
  object,
  x_axis,
  y_axis,
  x_order,
  y_order,
  color,
  width,
  height = 10
)
}
\arguments{
\item{object}{Seurat object}

\item{x_axis}{variable in meta data that is used for the y axis}

\item{y_axis}{variable in meta data that is used for the x axis}

\item{x_order}{vector determining the order of the x axis}

\item{y_order}{vector determining the order of the y axis}

\item{color}{color palette}

\item{width}{width of output plot (default: 10)}

\item{height}{height of output plot}
}
\value{
save stacked abundance barplot to folder \verb{/results/abundance}
}
\description{
create and save an abundance stacked bar plot in the folder \code{abundance}
}
\examples{
\dontrun{
stackedPlot(
object = sc_merge,
x_axis = "pool",
y_axis = "cluster",
x_order = unique(sc_merge$pool),
y_order = cluster_order,
color = col_vector,
width = 4)
}
}
