% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_cluster.R
\name{sub_cluster}
\alias{sub_cluster}
\title{Quick and dirty subclustering}
\usage{
sub_cluster(...)
}
\arguments{
\item{...}{Other arguments passed to
\link[scran:quickSubCluster]{quickSubCluster}
\itemize{
\item \code{x} A matrix of counts or log-normalized expression values (if
\code{normalize=FALSE}), where each row corresponds to a gene and each
column corresponds to a cell. Alternatively, a
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment} or
\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment} object
containing such a matrix.
\item \code{groups}: A vector of group assignments for all cells, usually
corresponding to cluster identities.
\item \code{normalize}: Logical scalar indicating whether each subset of \code{x}
should be log-transformed prior to further analysis.
\item \code{restricted}: Character vector containing the subset of groups in
\code{groups} to be subclustered. By default, all unique groups in
\code{groups} are used for subclustering, but this can be restricted to
specific groups of interest to save compute time.
\item \code{prepFUN}: A function that accepts a single
\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment} object
and returns another
\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment}
containing any additional elements required for clustering (e.g., PCA
results).
\item \code{min.ncells}: An integer scalar specifying the minimum number of cells in
a group to be considered for subclustering.
\item clusterFUN A function that accepts a single
\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment} object
and returns a vector of cluster assignments for each cell in that object.
\item \code{BLUSPARAM}: A \link[bluster:BlusterParam-class]{BlusterParam} object that is
used to specify the clustering via \link[bluster:clusterRows]{clusterRows}.
Only used when \code{clusterFUN=NULL}.
\item \code{assay.type}: String or integer scalar specifying the relevant assay.
}}
}
\value{
A factor
}
\description{
Performs a quick subclustering for all cells within each group.
}
\seealso{
\link[scran:quickSubCluster]{quickSubCluster}
}
