% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{theme_rect}
\alias{theme_rect}
\title{nice ggplot theme}
\usage{
theme_rect()
}
\description{
nice theme with square border
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(x = wt, y = mpg)) +
  geom_point()
p + theme_rect()
}
