% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.R
\name{top_deg}
\alias{top_deg}
\title{top_deg}
\usage{
top_deg(x, fdr = 0.01, lfc = 1, n = 10, group = "cluster")
}
\arguments{
\item{x}{data.frame or similar with DEGs.}

\item{fdr}{FDR used for filtering.}

\item{lfc}{log2 fold change used for filtering.}

\item{n}{number of genes per group to return.}

\item{group}{column name to group results.}
}
\description{
top_deg
}
