% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlnplot.R
\name{vlnPlotMisc}
\alias{vlnPlotMisc}
\title{Create a beautified violin plot for Seurat objects}
\usage{
vlnPlotMisc(
  object,
  features,
  cols = NULL,
  idents = NULL,
  pt.size = 0,
  sort = FALSE,
  assay = NULL,
  group.by = NULL,
  split.by = NULL,
  adjust = 1,
  y.max = NULL,
  same.y.lims = FALSE,
  log = FALSE,
  ncol = NULL,
  split.plot = FALSE,
  stack = FALSE,
  combine = TRUE,
  fill.by = "feature",
  flip = FALSE,
  add.noise = TRUE,
  raster = NULL,
  x.lab = NULL,
  y.lab = "Expression level",
  legend.position = "none",
  ...
)
}
\arguments{
\item{object}{A Seurat object}

\item{features}{Features to plot (e.g. genes)}

\item{cols}{Colors for the plot (optional)}

\item{idents}{Identity classes to include in the plot}

\item{pt.size}{Size of the points in the plot (default: 0)}

\item{sort}{Whether to sort the identity classes (default: FALSE)}

\item{assay}{Assay to use}

\item{group.by}{Variable to group cells by}

\item{split.by}{Variable to split plots by}

\item{adjust}{Adjust parameter for density estimation (default: 1)}

\item{y.max}{Maximum y-axis value}

\item{same.y.lims}{Use the same y-axis limits for all plots (default: FALSE)}

\item{log}{Use log scale for y-axis (default: FALSE)}

\item{ncol}{Number of columns in the plot grid}

\item{split.plot}{Whether to split the plot (default: FALSE)}

\item{stack}{Whether to stack the plots (default: FALSE)}

\item{combine}{Whether to combine the plots (default: TRUE)}

\item{fill.by}{How to fill the violin plots (default: "feature")}

\item{flip}{Flip the plot coordinates (default: FALSE)}

\item{add.noise}{Add jitter to the plot (default: TRUE)}

\item{raster}{Convert points to raster format}

\item{x.lab}{Label for x-axis (optional)}

\item{y.lab}{Label for y-axis (default: "Expression level")}

\item{legend.position}{Position of the legend (default: "none")}

\item{...}{Additional arguments passed to Seurat's VlnPlot function}
}
\value{
A ggplot object representing the beautified violin plot
}
\description{
This function enhances Seurat's VlnPlot by adding mean points, error bars, and custom theming.
}
\examples{
\dontrun{
VlnPlotMisc(object = seurat.data, features = c("CD79A", "MS4A1"), group.by = "celltype",pt.size = 0)
}

}
