% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{abBoxPlot}
\alias{abBoxPlot}
\title{abundance boxplot plot}
\usage{
abBoxPlot(
  object,
  cluster_idents,
  sample,
  cluster_order,
  group_by,
  group_order,
  color,
  width = 9,
  height = ceiling(length(unique(object@meta.data[[cluster_idents]]))/4) * 3,
  paired = FALSE,
  number_of_tests
)
}
\arguments{
\item{object}{Seurat object}

\item{cluster_idents}{variable in meta data with cluster names}

\item{sample}{variable in meta data for each sample}

\item{cluster_order}{vector determining the order of the clusters}

\item{group_by}{variable in meta data that categorize samples in groups}

\item{group_order}{vector determining the order of the samples}

\item{color}{color palette}

\item{width}{width of output plot (default: 9)}

\item{height}{height of output plot (default: length of cluster_idents divided by four, ceiling, times three)}

\item{paired}{logical indicating whether you want a paired test (default FALSE)}

\item{number_of_tests}{number of tests to be performed}
}
\value{
save abundance box plot in the folder \verb{/results/abundance}
}
\description{
create and save an abundance boxplot in the folder \code{abundance}
}
\examples{
\dontrun{
abBoxPlot(object = aie_pbmc,
         cluster_idents = "cluster",
         sample = "sample",
       cluster_order = cluster_order,
         group_by = "AIE_type",
         group_order = c("control", "CASPR2", "LGI1"),
         color = my_cols)
}
}
