% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_helper.R
\name{abundanceTbl}
\alias{abundanceTbl}
\title{wrapper function to calculate absolute and relative abundance}
\usage{
abundanceTbl(object, row_var, col_var)
}
\arguments{
\item{object}{Seurat object}

\item{row_var}{variable in meta data that will represent the rows}

\item{col_var}{variable in meta data that will represent the columns}
}
\description{
wrapper function to calculate absolute and relative abundance
}
\examples{
\dontrun{abundanceTbl(object = aie, row_var = "predicted.id", col_var = "AIE_type")}
}
