% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{abVolPlot}
\alias{abVolPlot}
\title{abundance volcano plot}
\usage{
abVolPlot(
  object,
  cluster_idents,
  sample,
  cluster_order,
  group_by,
  group1,
  group2,
  color,
  width = 5,
  height = 5,
  min_cells = 10,
  paired = FALSE,
  dir_output = "."
)
}
\arguments{
\item{object}{Seurat object}

\item{cluster_idents}{variable in meta data with cluster names}

\item{sample}{variable in meta data for each sample}

\item{cluster_order}{vector determining the order of the clusters}

\item{group_by}{variable in meta data that categorize samples in groups}

\item{group1}{first group (nominator)}

\item{group2}{second group (denominator)}

\item{color}{color palette}

\item{width}{width of output plot (default: 5)}

\item{height}{height of output plot (default: 5)}

\item{min_cells}{remove all clusters that have less than minimal amount of cells (default = 10)}

\item{paired}{logical indicating whether you want a paired test (default FALSE)}

\item{dir_output}{directory to save the output plot (default: ".")}
}
\value{
save volcano abundance plot
}
\description{
create and save an abundance volcano bar plot in the folder \code{abundance}
}
\examples{
library(Seurat)
library(rstatix)
set.seed(123)
pbmc_small$predicted.id <- sample(c("Cluster1", "Cluster2"), ncol(pbmc_small), replace = TRUE)
pbmc_small$sample <- sample(c("Sample1", "Sample2"), ncol(pbmc_small), replace = TRUE)
pbmc_small$AIE_type <- sample(c("LGI1", "control"), ncol(pbmc_small), replace = TRUE)
suppressWarnings(
    plot <- abVolPlot(
        object = pbmc_small,
        cluster_idents = "predicted.id",
        sample = "sample",
        cluster_order = c("Cluster1", "Cluster2"),
        group_by = "AIE_type",
        group1 = "LGI1",
        group2 = "control",
        color = c("Cluster1" = "blue", "Cluster2" = "red"),
        width = 5,
        height = 5,
        dir_output = "."
    )
)
unlink("volcano_plot_predicted.id_pbmc_small_LGI1_control.pdf")
}
