% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_helper.R
\name{avgExp}
\alias{avgExp}
\title{wrapper function for Seurats AverageExpression}
\usage{
avgExp(par, object, assay, slot, ortho = "none")
}
\arguments{
\item{par}{column name in markers.csv}

\item{object}{Seurat object}

\item{assay}{which assay to use}

\item{slot}{which slot to use}

\item{ortho}{convert to orthologues? Allowed values: \code{none}, \code{mouse2human} or \code{human2mouse}}
}
\value{
matrix with genes as rows and identity clases as columns
}
\description{
create average expression matrix with Seurat based on \code{markers.csv}
}
\examples{
library(Seurat)
markers <- data.frame(Bc = c("CD19", "MS4A1", "CD79B"))
write.csv(markers, "markers.csv")
avgExp("Bc", object = pbmc_small, assay = "RNA", slot = "data")
unlink("markers.csv") 
}
