% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{calculate_module_scores}
\alias{calculate_module_scores}
\alias{calculate_module_scores.Seurat}
\title{calculate_module_scores}
\usage{
calculate_module_scores(x)

calculate_module_scores.Seurat(
  x,
  gene_modules,
  group.by = "seurat_clusters",
  assay = NULL,
  slot = "data"
)
}
\arguments{
\item{x}{object with gene expression levels per cell.}

\item{gene_modules}{data.frame with columns gene and module.}

\item{group.by}{identity to use for grouping cells (default: seurat_clusters).}

\item{assay}{assay to use for Seurat objects.}

\item{slot}{slot to use for Seurat objects.}

\item{size}{size of the points.}

\item{...}{arguments passed down to methods.}
}
\description{
Calculate gene module score per cell clusters.
}
