% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{compStat}
\alias{compStat}
\title{significance for boxplot}
\usage{
compStat(x_var, group, data, paired)
}
\arguments{
\item{x_var}{numeric variable (x in the formula \code{x ~ group})}

\item{group}{grouping variable (group in the formula \code{x ~ group})}

\item{data}{data frame containing the variables}

\item{paired}{logical value. Do you want to do perform pair testing?}
}
\value{
data frame with adjusted signficance values
}
\description{
calculcate significance for ggsignif
}
\examples{
\dontrun{compStat(x_var = "pct", group = "type", data = bp_data)}
}
