% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{create_training_sets}
\alias{create_training_sets}
\alias{create_training_sets.SingleCellExperiment}
\title{create_training_sets}
\usage{
create_training_sets(x, ...)

\method{create_training_sets}{SingleCellExperiment}(x, frac = 0.1)
}
\arguments{
\item{x}{SingleCellExperiment object.}

\item{frac}{fraction of dataset for testing.}
}
\value{
list with two SingleCellExperiment, one with train and other with test data.
}
\description{
Divides a SingleCellExperiment object into train/test subsets.
}
