% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{dotplotPropeller}
\alias{dotplotPropeller}
\title{plot propeller results in a barplot}
\usage{
dotplotPropeller(data, color, filename, width = 5, height = 5)
}
\arguments{
\item{data}{A dataframe containing the results from propeller calculation}

\item{color}{A vector of colors for the clusters in the plot}

\item{filename}{A character representing the file name of the plot}

\item{width}{The width of the plot}

\item{height}{The height of the plot}
}
\value{
save propeller abundance barplot
}
\description{
The function creates a barplot of the propeller results and saves the plot in results/abundance folder
}
\examples{
\dontrun{dotplotPropeller(data = pnp_ctrl_csf_sex_age, color = cluster_col, filename = "pnp_ctrl_csf_sex_age")}
}
