% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_dots.R
\name{facet_dots}
\alias{facet_dots}
\alias{facet_dots,ANY-method}
\alias{facet_dots,SingleCellExperiment-method}
\title{Create a facet dot plot of expression values}
\usage{
facet_dots(x, ...)

\S4method{facet_dots}{ANY}(
  x,
  marker_list,
  clusters,
  cluster2cell = NULL,
  flip = TRUE,
  facet_args = list(scales = "free", space = "free"),
  ...
)

\S4method{facet_dots}{SingleCellExperiment}(x, ..., clusters = NULL)
}
\arguments{
\item{x}{Currently, only
\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment} object
is supported.}

\item{...}{
  Arguments passed on to \code{\link[scater:plotDots]{scater::plotDots}}
  \describe{
    \item{\code{block}}{String specifying the field of \code{\link{colData}(object)} containing a blocking factor (e.g., batch of origin).
Alternatively, any value that can be used in the \code{by} argument to \code{\link[scater]{retrieveCellInfo}}.}
    \item{\code{exprs_values}}{A string or integer scalar indicating which assay of \code{object} should be used as expression values.}
    \item{\code{detection_limit}}{Numeric scalar providing the value above which observations are deemed to be expressed.}
    \item{\code{zlim}}{A numeric vector of length 2, specifying the upper and lower bounds for color mapping of expression values.
Values outside this range are set to the most extreme color.
If \code{NULL}, it defaults to the range of the expression matrix.
If \code{center=TRUE}, this defaults to the range of the centered expression matrix, made symmetric around zero.}
    \item{\code{color}}{A vector of colours specifying the palette to use for increasing expression. 
This defaults to \link[viridis]{viridis} if \code{center=FALSE}, and the the \code{"RdYlBu"}
color palette from \code{\link[RColorBrewer]{brewer.pal}} otherwise.}
    \item{\code{max_detected}}{Numeric value specifying the cap on the proportion of 
detected expression values.}
    \item{\code{by_exprs_values}}{A string or integer scalar specifying which assay to obtain expression values from, for entries of \code{other_fields}.}
    \item{\code{swap_rownames}}{Column name of \code{rowData(object)} to be used to 
identify features instead of \code{rownames(object)} when labelling plot 
elements.}
    \item{\code{center}}{A logical scalar indicating whether each feature should have its mean expression 
(specifically, the mean of averages across all groups) centered at zero prior to plotting.}
    \item{\code{scale}}{A logical scalar specifying whether each row should have its
average expression values scaled to unit variance prior to plotting.}
    \item{\code{low_color,high_color,max_ave}}{Deprecated arguments.}
  }}

\item{marker_list}{A named list contaning the markers for each cell types.
names indicate the cell type label and values indicate the markers for this
cell type.}

\item{clusters}{A factor (or vector coercible into a factor) specifying the
group to which each cell in \code{x} belongs. Alternatively, if \code{x} is a
\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment}, e.g,
String specifying the field of \code{colData(x)} containing the grouping factor.
In this way, if clusters is \code{NULL}, "label" in \code{colData(x)} will be
extracted.}

\item{cluster2cell}{A named character or factor returned by
\code{\link[=annotate_clusters]{annotate_clusters()}}.}

\item{flip}{A scalar logical indicates whether flipping the plot.}

\item{facet_args}{A named arguments list passed to
\link[ggplot2:facet_grid]{facet_grid}.}
}
\value{
A ggplot2 object.
}
\description{
Create a dot plot of expression values for a grouping of cells, where the
size and color of each dot represents the proportion of detected expression
values and the average expression, respectively, for each feature in each
group of cells.
}
