% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{fPlot}
\alias{fPlot}
\title{Seurat feature plot}
\usage{
fPlot(
  path,
  object,
  par,
  reduction,
  width = 16,
  height = ceiling(length(genes_found)/4) * 3,
  order,
  dir_output = "."
)
}
\arguments{
\item{path}{path to markers.csv}

\item{object}{Seurat object}

\item{par}{column name in markers.csv}

\item{reduction}{a character string specifying the dimension reduction}

\item{width}{width of output plot (default: 16)}

\item{height}{height of output plot (default: length of genes divided by four, ceiling, times three)}

\item{order}{should the feature plot be ordered in order of expression}

\item{dir_output}{directory to save the output plot (default: ".")}
}
\value{
save feature plot to folder \verb{/results/featureplot/}
}
\description{
create and save a nice Seurat feature plot in folder \code{featureplot}
}
\examples{
library(Seurat)
markers <- data.frame(B = c("MS4A1", "CD79A"))
write.csv(markers, "markers.csv")
fPlot(path = "markers.csv", object = pbmc_small, par = "B", reduction = "tsne", order = TRUE, dir_output = ".")
unlink("markers.csv")
unlink("fp_pbmc_small_B.png")
}
