% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_coldata}
\alias{get_coldata}
\alias{get_coldata.SingleCellExperiment}
\alias{get_coldata.seurat}
\title{get_coldata}
\usage{
get_coldata(x)

\method{get_coldata}{SingleCellExperiment}(x)

\method{get_coldata}{seurat}(x)
}
\arguments{
\item{x}{object.}
}
\description{
get_coldata
}
