% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{get_density}
\alias{get_density}
\title{get_density}
\usage{
get_density(x, y, n = 100)
}
\arguments{
\item{x}{x coordinate of data}

\item{y}{y coordinate of data}

\item{n}{Number of grid points in each direction. Can be scalar or a length-2 integer vector.}
}
\description{
get_density
}
\note{
I took this implementation from http://slowkow.com/notes/ggplot2-color-by-density/
}
