% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_rowdata}
\alias{get_rowdata}
\alias{get_rowdata.SingleCellExperiment}
\alias{get_rowdata.seurat}
\title{get_rowdata}
\usage{
get_rowdata(x)

\method{get_rowdata}{SingleCellExperiment}(x)

\method{get_rowdata}{seurat}(x)
}
\arguments{
\item{x}{object.}
}
\description{
get_rowdata
}
