% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helper.R
\name{lss}
\alias{lss}
\title{pretty function to list environment objects}
\usage{
lss()
}
\value{
tibble with environment objects arranged by size
}
\description{
list objects from the environment, changes to human-readable units and arranges them by size
}
