% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{ModulePlot}
\alias{ModulePlot}
\title{module plot}
\usage{
ModulePlot(x_var, module, object, color)
}
\arguments{
\item{x_var}{variable in meta data which represents the x-axis}

\item{module}{variable in meta data of the module score}

\item{object}{Seurat object}

\item{color}{color palette}
}
\value{
plot module plot
}
\description{
create combined violin and boxplot of the module score
}
\examples{
library(Seurat)
set.seed(123)
pbmc_small$AIE_type <- sample(c("control", "CASPR2", "LGI1"), ncol(pbmc_small), replace = TRUE)
module1 <- list(c(rownames(pbmc_small)[1:100]))
pbmc_small <- AddModuleScore(pbmc_small, features = module1, assay = "RNA", name = "module", ctrl = 5)
ModulePlot(
  x_var = "AIE_type",
  module = "module1",
  object = pbmc_small,
  color = c("control" = "blue", "CASPR2" = "red", "LGI1" = "green")
)
}
