% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-palettes.R
\name{pal_scanpy}
\alias{pal_scanpy}
\title{SCANPY Color Palettes}
\usage{
pal_scanpy(palette = c("zeileis", "godsnot"), alpha = 1)
}
\arguments{
\item{palette}{Palette type.
Currently there is one available option: \code{"default"} (102-color palette).}

\item{alpha}{Transparency level, a real number in \verb{(0, 1]}.
See \code{alpha} in \link[grDevices:rgb]{rgb} for details.}
}
\description{
Color palette inspired by plots in
\emph{The New England Journal of Medicine}.
}
