% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{pcaSeurat}
\alias{pcaSeurat}
\title{plot PCA of cluster abundances}
\usage{
pcaSeurat(object, cluster, sample, condition, width = 20, height = 5)
}
\arguments{
\item{object}{A Seurat object}

\item{cluster}{A character string indicating the cluster column in the metadata of the Seurat object}

\item{sample}{A character string indicating the sample column in the metadata of the Seurat object}

\item{condition}{A character string indicating the condition column in the metadata of the Seurat object}

\item{width}{The width of the plot}

\item{height}{The height of the plot}
}
\value{
A ggplot object.
}
\description{
The function creates a PCA plot of the cluster abundances and saves the plot in results/abundance folder
}
\examples{
\dontrun{
pcaSeurat(
  object = sc_final,
  cluster = "cluster",
  sample = "patient",
  condition = "condition")
}
}
