% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_coord}
\alias{plot_coord}
\alias{plot_coord.seurat}
\alias{plot_coord.Seurat}
\alias{plot_coord.SingleCellExperiment}
\alias{plot_coord.data.frame}
\title{plot_coord}
\usage{
plot_coord(x, ...)

\method{plot_coord}{seurat}(x, size = 0.1, color = NULL,
  shape = NULL, label = NULL, expand = NULL, ...)

\method{plot_coord}{Seurat}(x, size = 0.1, color = NULL,
  shape = NULL, label = NULL, expand = NULL, ...)

\method{plot_coord}{SingleCellExperiment}(x, size = 0.1, color = NULL,
  shape = NULL, label = NULL, expand = NULL, ...)

\method{plot_coord}{data.frame}(x, size = 0.1, color = NULL,
  shape = NULL, label = NULL, expand = NULL, ...)
}
\arguments{
\item{x}{object from which to plot coordinates.}

\item{...}{further arguments passed down to get_coord().}

\item{size}{size of points in geom_point()}

\item{color}{color or column to map to color..}

\item{shape}{shape or column to map to shape.}

\item{label}{name of column used to label plot (e.g. clusters).}

\item{expand}{logical; whether to expand one column to show presence/absence.}
}
\description{
plot_coord
}
