% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpdb.R
\name{plot_cpdb_dotplot}
\alias{plot_cpdb_dotplot}
\title{plot_cpdb_dotplot}
\usage{
plot_cpdb_dotplot(
  x,
  cutoff = 0.001,
  filter = NULL,
  cells = NULL,
  proteins = NULL,
  cluster_cells = FALSE,
  cluster_proteins = FALSE
)
}
\arguments{
\item{cutoff}{cutoff for pvalues.}

\item{filter}{filter for cell pairs using regular expression.}

\item{cells}{subset cell pairs.}

\item{proteins}{subset protein pairs.}

\item{cluster_cells}{logical; whether to cluster cells.}

\item{cluster_proteins}{logigcal; whether to cluster proteins.}

\item{path}{path to the folder containing cellphonedb results.}
}
\description{
Plot cellphonedb results as a dotplot.
}
