% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diffusion_map.R
\name{plot_diffusion_map}
\alias{plot_diffusion_map}
\alias{plot_diffusion_map,SingleCellExperiment,DiffusionMap-method}
\alias{plot_diffusion_map,SingleCellExperiment,DPT-method}
\title{Plot the diffusion map}
\usage{
plot_diffusion_map(x, y, ...)

\S4method{plot_diffusion_map}{SingleCellExperiment,DiffusionMap}(x, y, ...)

\S4method{plot_diffusion_map}{SingleCellExperiment,DPT}(
  x,
  y,
  dimred = "DiffusionMap",
  colour_by = "DPT",
  add_paths = TRUE,
  root = NULL,
  paths_to = NULL,
  path_args = list(),
  w_width = 0.1,
  ...,
  ncomponents = 2L
)
}
\arguments{
\item{x}{A \link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment}
object.}

\item{y}{A \link[destiny:DiffusionMap-class]{DiffusionMap} or \link[destiny:DPT]{DPT}
object.}

\item{...}{Other arguments passed to
\link[scater:plotReducedDim]{plotReducedDim}.}

\item{dimred}{A string or integer scalar indicating the reduced dimension
result in \code{reducedDims(x)} to plot.}

\item{colour_by}{Specification of a column metadata field or a feature to
colour by, see the by argument in
\code{\link[scater:retrieveCellInfo]{?retrieveCellInfo}} for possible values.}

\item{add_paths}{A scalar logical value indicates whether to add path.}

\item{root}{Root branch ID. Will be used as the start of the DPT. (default:
lowest branch ID).}

\item{paths_to}{Numeric Branch IDs. Are used as target(s) for the path(s) to
draw.}

\item{path_args}{A list of arguments to passed to
\link[ggplot2:geom_path]{geom_path}.}

\item{w_width}{Window width for smoothing the path (see
\link[smoother:smth.gaussian]{smth.gaussian}).}

\item{ncomponents}{A numeric scalar indicating the number of dimensions to
plot, starting from the first dimension. Alternatively, a numeric vector
specifying the dimensions to be plotted.}
}
\value{
A ggplot2 object depicting the Diffusion map, if add_paths is \code{TRUE},
DPT of the root is also added in this map.
}
\description{
Plot the diffusion map
}
