% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_pairs}
\alias{plot_pairs}
\alias{plot_pairs.Seurat}
\title{plot_pairs}
\usage{
plot_pairs(x, ...)

\method{plot_pairs}{Seurat}(
  x,
  features = NULL,
  slot = "data",
  assay = NULL,
  color.by = NULL,
  add.contour = FALSE,
  color.contour = "red",
  ...
)
}
\arguments{
\item{x}{object to plot.}

\item{...}{arguments passed down to methods.}

\item{features}{features to use for paired plots. If null all features in assay.}

\item{slot}{slot from Seurat object.}

\item{assay}{assay from Seurat object.}

\item{color.by}{column from metadata to use for coloring points.}

\item{add.contour}{whether to add contour plots.}

\item{color.contour}{color for contour lines.}
}
\description{
plot_pairs
}
