% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_purity}
\alias{plot_purity}
\alias{plot_purity.SingleCellExperiment}
\alias{plot_purity.DataFrame}
\alias{plot_purity.data.frame}
\title{plot_purity}
\usage{
plot_purity(x, ...)

\method{plot_purity}{SingleCellExperiment}(x, ...)

\method{plot_purity}{DataFrame}(x, ...)

\method{plot_purity}{data.frame}(x, col.x = "celltype",
  col.y = "cluster", label = FALSE, ...)
}
\arguments{
\item{x}{a suitable object.}

\item{...}{parameters passed down to methods.}

\item{col.x}{name of column to be use in x-axis.}

\item{col.y}{name of column to be use in y-axis.}
}
\description{
plot_purity
}
