% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_volcano}
\alias{plot_volcano}
\title{plot_volcano}
\usage{
plot_volcano(
  x,
  group.by = "cluster",
  groups = NULL,
  n = 10,
  fdr = 0.01,
  lfc = 1,
  min.segment.length = 0,
  max.overlaps = Inf,
  ...
)
}
\arguments{
\item{x}{data.frame with DEG results from FindAllMarkers.}

\item{group.by}{column to group plots by.}

\item{groups}{clusters to plot. If null, plot all.}

\item{n}{number of genes (per down/up-regulated) to highlight.}

\item{fdr}{false discovery rate cutoff to highlight genes.}

\item{lfc}{log fold change cutoff to highlight genes.}

\item{min.segment.length}{minimum segment length for ggrepel.}

\item{...}{arguments passed down to ggrepel functions.}

\item{max.overkaps}{max number of overlaps for ggrepel.}
}
\description{
plot_volcano
}
