% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotEnrichr}
\alias{plotEnrichr}
\title{enrichr plot}
\usage{
plotEnrichr(filename, sheet, width, height, dir_input = ".", dir_output = ".")
}
\arguments{
\item{filename}{name of .xlsx file without extension and without \code{enrichr_}}

\item{sheet}{name of the excel sheet}

\item{width}{width of output plot}

\item{height}{height of output plot}

\item{dir_input}{directory where the input file (.xlsx) is located (default: ".")}

\item{dir_output}{directory to save the output plot (default: ".")}
}
\value{
create and save enrichr plot
}
\description{
plot enrichr results in a bar plot
}
