% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotEnrichr}
\alias{plotEnrichr}
\title{enrichr plot}
\usage{
plotEnrichr(filename, sheet, width, height)
}
\arguments{
\item{filename}{name of file without extension and without \code{enrichr_}}

\item{sheet}{name of the excel sheet}

\item{width}{width of output plot}

\item{height}{height of output plot}
}
\value{
save enrichr plot
}
\description{
plot enrichr results in a bar plot
}
\examples{
\dontrun{plotEnrichr("de_ALZ_Naive_CSF_neg_pDC", sheet = "GO_Biological_Process_2021", width = 10, height = 5)}
}
