% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotSlingshot}
\alias{plotSlingshot}
\title{plot slingshot results}
\usage{
plotSlingshot(object, lineage, pt, curves)
}
\arguments{
\item{object}{A Seurat object}

\item{lineage}{A character string indicating the lineage of interest.}

\item{pt}{A numeric vector of pseudotime values.}

\item{curves}{A dataframe containing the curves.}
}
\value{
A ggplot object.
}
\description{
The function creates a colored UMAP plot and curves split by lineage
}
\examples{
library(Seurat)
set.seed(123)
pbmc_small$lineage <- sample(c("Lineage1", "Lineage2"), ncol(pbmc_small), replace = TRUE)
pbmc_small$umap <-
CreateDimReducObject(
 embeddings =
   Embeddings(
     pbmc_small,
     reduction = "tsne"
   ),
 key = "UMAP_",
 assay = "RNA"
)
curves <- data.frame(
    UMAP_1 = runif(ncol(pbmc_small), min = -10, max = 10),
    UMAP_2 = runif(ncol(pbmc_small), min = -10, max = 10),
    Lineage = sample(c("Lineage1", "Lineage2"), ncol(pbmc_small), replace = TRUE)
)
pt <- matrix(runif(ncol(pbmc_small) * 2), ncol = 2)
colnames(pt) <- c("Lineage1", "Lineage2")
plot <- plotSlingshot(object = pbmc_small, lineage = "Lineage1", pt = pt)
}
