% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.R
\name{reduce_dim}
\alias{reduce_dim}
\alias{reduce_dim.SingleCellExperiment}
\alias{reduce_dim.matrix}
\title{reduce_dim}
\usage{
reduce_dim(x, method = "pca", dims = 2, ...)

\method{reduce_dim}{SingleCellExperiment}(
  x,
  method = "pca",
  dims = 2,
  assay.name = "logcounts",
  coord.name = method,
  perplexity = NULL,
  initial_dims = 50,
  ...
)

\method{reduce_dim}{matrix}(
  x,
  method = "pca",
  dims = 2,
  perplexity = 30,
  initial_dims = 50,
  ...
)
}
\arguments{
\item{x}{matrix object.}

\item{method}{tyoe of dimensionality reduction: pca, tsne, umap.}

\item{dims}{number of dimensions to keep in final result.}

\item{...}{arguments to be passed down to methods.}

\item{assay.name}{name of assay slot.}

\item{coord.name}{name of reducedDim slot.}

\item{perplexity}{perplexity for tSNE.}

\item{initial_dims}{initial PCA dimensions for tSNE.}
}
\description{
reduce_dim
}
