% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_diffusion_map.R
\name{run_diffusion_map}
\alias{run_diffusion_map}
\alias{run_diffusion_map,ANY-method}
\alias{run_diffusion_map,SingleCellExperiment-method}
\title{Create a diffusion map of cells}
\usage{
run_diffusion_map(x, ...)

\S4method{run_diffusion_map}{ANY}(x, root = NULL, ..., w_width = 0.1, dpt = TRUE, size = 1L)

\S4method{run_diffusion_map}{SingleCellExperiment}(
  x,
  dimred = "PCA",
  ...,
  n_dimred = NULL,
  assay.type = "logcounts"
)
}
\arguments{
\item{x}{A numeric matrix of counts with cells in columns and features in
rows.

Alternatively, a
\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment} object
containing such a matrix.}

\item{...}{Other arguments passed to \link[destiny:DiffusionMap-class]{DiffusionMap}.}

\item{root}{The cell index/indices from which to calculate the DPT(s)
(integer of length 1-3). If length more than \code{size}, we will sample 3
elements from it. Can be also a logical atomic value, in this way, we will
sample the true integer index.}

\item{w_width}{Window width to use for deciding the branch cutoff.}

\item{dpt}{A scalar logical value indicates whether to run
\link[destiny:DPT]{DPT}.}

\item{size}{A scalar numeric ranges from 1 to 3 specifying how many roots
should be used.}

\item{dimred}{String or integer scalar specifying the existing dimensionality
reduction results to use.}

\item{n_dimred}{Integer scalar or vector specifying the dimensions to use if
dimred is specified.}

\item{assay.type}{A string or integer scalar indicating which \code{assays} in the
\code{x} contains the count matrix. Only used when dimred is \code{NULL}.}
}
\value{
A list or a modified x that contains the DiffusionMap coordinates in
\code{reducedDim(x, "DiffusionMap")}, if \code{dpt} is TRUE, "branch" and "dpt" data
are also kept in \code{colData(x)} named as "destiny_branch" and "destiny_dpt"
respectively.
}
\description{
A modified version of \link[destiny:DiffusionMap-class]{DiffusionMap}, this function
will add the DiffusionMap results components into x if it is a
\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment} object.
}
