% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_diffusion_map.R
\name{run_diffusion_map}
\alias{run_diffusion_map}
\alias{run_diffusion_map,ANY-method}
\alias{run_diffusion_map,DiffusionMap-method}
\alias{run_diffusion_map,SingleCellExperiment-method}
\title{Create a diffusion map of cells}
\usage{
run_diffusion_map(x, ...)

\S4method{run_diffusion_map}{ANY}(x, root = NULL, ..., w_width = 0.1, dpt = TRUE, size = 1L)

\S4method{run_diffusion_map}{DiffusionMap}(x, root = NULL, ..., w_width = 0.1, dpt = TRUE, size = 1L)

\S4method{run_diffusion_map}{SingleCellExperiment}(
  x,
  dimred = "PCA",
  ...,
  n_dimred = NULL,
  assay.type = "logcounts"
)
}
\arguments{
\item{x}{A numeric matrix of counts with cells in columns and features in
rows.

Alternatively, a
\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment} object
containing such a matrix, or a \link[destiny:DiffusionMap-class]{DiffusionMap}
object.}

\item{...}{Other arguments passed to \link[destiny:DiffusionMap-class]{DiffusionMap}.}

\item{root}{The cell index/indices from which to calculate the DPT(s)
(integer of length \code{1-3}). If the length of root is more than \code{size} or  a
logical atomic vector with the compatible length of \code{x}(the same length of
\code{ncol(x)} if x is a simple matrix or
\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment} or the
same length of \verb{nrow(}\link[destiny:destiny-generics]{eigenvectors}\verb{(x))} if x is a
\link[destiny:DiffusionMap-class]{DiffusionMap} object), we will index with root
value and choose the top \code{size} elements with the maximal
\link[destiny:destiny-generics]{eigenvectors} CD1 values.}

\item{w_width}{Window width to use for deciding the branch cutoff.}

\item{dpt}{A scalar logical value indicates whether to run
\link[destiny:DPT]{DPT}.}

\item{size}{A scalar integer ranges from 1 to 3 specifying how many roots
should be used.}

\item{dimred}{String or integer scalar specifying the existing dimensionality
reduction results to use.}

\item{n_dimred}{Integer scalar or vector specifying the dimensions to use if
dimred is specified.}

\item{assay.type}{A string or integer scalar indicating which \code{assays} in the
\code{x} contains the count matrix. Only used when dimred is \code{NULL}.}
}
\value{
A \link[destiny:DiffusionMap-class]{DiffusionMap} or \link[destiny:DPT]{DPT}
object depend on \code{dpt}.
}
\description{
A modified version of \link[destiny:DiffusionMap-class]{DiffusionMap} and
\link[destiny:DPT]{DPT}.
}
