% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample_cells}
\alias{sample_cells}
\alias{sample_cells.SingleCellExperiment}
\title{sample_cells}
\usage{
sample_cells(x, ...)

\method{sample_cells}{SingleCellExperiment}(x, group = NULL, n = 20,
  seed = NULL, ...)
}
\arguments{
\item{x}{an object to sample from.}

\item{...}{parameters passed down to methods.}

\item{group}{optional grouping variable.}

\item{n}{number of cells to sample (per group).}

\item{seed}{seed for the random number generation.}
}
\description{
Subset a single cell experiment object by sampling the number of specified cells,
optionally by grouping.
}
