% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmisc-package.R
\docType{package}
\name{scmisc-package}
\alias{scmisc}
\alias{scmisc-package}
\title{scmisc: Miscellaneus tools for the analysis of single cell data}
\description{
Miscellaneus tools for the analysis of single cell data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ddiez/scmisc}
  \item Report bugs at \url{https://github.com/ddiez/scmisc/issues}
}

}
\author{
\strong{Maintainer}: Diego Diez \email{diego10ruiz@gmail.com} (\href{https://orcid.org/0000-0002-2325-4893}{ORCID})

}
\keyword{internal}
