% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_markers.R
\name{score_markers}
\alias{score_markers}
\alias{score_markers,ANY-method}
\alias{score_markers,SingleCellExperiment-method}
\alias{score_markers,SummarizedExperiment-method}
\title{Score marker genes}
\usage{
score_markers(x, ...)

\S4method{score_markers}{ANY}(
  x,
  restricted = NULL,
  top_n = 20L,
  order_by = "mean.AUC",
  order = -1L,
  cellmarker = FALSE,
  clusters = NULL,
  lfc = 1L,
  features = NULL,
  ...
)

\S4method{score_markers}{SingleCellExperiment}(
  x,
  ...,
  clusters = NULL,
  features = NULL,
  assay.type = "logcounts"
)

\S4method{score_markers}{SummarizedExperiment}(x, ..., clusters = NULL, assay.type = "logcounts")
}
\arguments{
\item{x}{A matrix-like object containing log-normalized expression values,
with genes in rows and cells in columns. Alternatively, a
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment} or
\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment} object
containing such a matrix in its assays.}

\item{...}{For the generic, further arguments to pass to specific methods.}

\item{restricted}{An atomic vector coerced to characters containing the
subset of groups in clusters to be returned. By default, all unique groups in
clusters will be returned, but this can be restricted to specific groups of
interest.}

\item{top_n}{A scalar integer indicates derive how many top features from the
results. If cellmarker is \code{TRUE}, these features were then searched from
cellmarker database.}

\item{order_by}{A character vector of column names by which to order. By
default, sorts over \code{mean.AUC}. See \link[scran:scoreMarkers]{scoreMarkers} for
full summary scores names.}

\item{order}{An integer vector with only possible values of \code{1}and \code{-1},
corresponding to ascending and descending order. The length of order must be
either 1 or equal to that of \code{order_by}. If \code{length(order) == 1}, it is
recycled to \code{length(order_by)}.}

\item{cellmarker}{A scalar logical indicates if search the CellMarker
database use the \code{top_n} genes. See \code{\link[=cellmarker_search]{cellmarker_search()}}}

\item{clusters}{A factor (or vector coercible into a factor) specifying the
group to which each cell in \code{x} belongs. Alternatively, String specifying
the field of \code{colData(x)} containing the grouping factor if \code{x} is a
\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment}, e.g.,
usually clusters. See \link[scran:scoreMarkers]{scoreMarkers} groups argument.}

\item{lfc}{A numeric scalar specifying the log-fold change threshold to
compute effect sizes against.}

\item{features}{This can be a logical, integer or character vector indicating
the rows of x to use. See \link[scran:scoreMarkers]{scoreMarkers} subset.row
argument}

\item{assay.type}{A string or integer scalar indicating which \code{assays} in the
\code{x} contains the count matrix.}
}
\value{
A data.frame
}
\description{
Compute various summary scores for potential marker genes to
distinguish between groups of cells.
}
