% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort.R
\name{sort_features}
\alias{sort_features}
\alias{sort_features.Seurat}
\alias{sort_features.SingleCellExperiment}
\alias{sort_features.matrix}
\title{sort_features}
\usage{
sort_features(x, ...)

\method{sort_features}{Seurat}(x, features = NULL, slot = "scale.data",
  assay = NULL, ...)

\method{sort_features}{SingleCellExperiment}(x, features = NULL,
  assay = NULL, ...)

\method{sort_features}{matrix}(x, ...)
}
\arguments{
\item{x}{an object with features to sort.}

\item{...}{rguments passed down to methods.}

\item{features}{a character vector of features to sort.}

\item{slot}{the slot to obtain the data from.}

\item{assay}{the assay to obtain the data from.}
}
\description{
Sort features (typically genes) using Pearson correlation.
}
